@@ -43,7 +43,7 @@ public final class NIOHTTP2Handler: ChannelDuplexHandler {
     private static let clientMagic: StaticString = "PRI * HTTP/2.0\r\n\r\nSM\r\n\r\n"
 
     /// The event loop on which this handler will do work.
-    private let eventLoop: EventLoop?
+    @usableFromInline internal let _eventLoop: EventLoop?
 
     /// The connection state machine. We always have one of these.
     private var stateMachine: HTTP2ConnectionStateMachine
@@ -76,7 +76,7 @@ public final class NIOHTTP2Handler: ChannelDuplexHandler {
     private var wroteFrame: Bool = false
 
     /// This object deploys heuristics to attempt to detect denial of service attacks.
-    private var denialOfServiceValidator: DOSHeuristics
+    private var denialOfServiceValidator: DOSHeuristics<RealNIODeadlineClock>
 
     /// The mode this handler is operating in.
     private let mode: ParserMode
@@ -106,6 +106,8 @@ public final class NIOHTTP2Handler: ChannelDuplexHandler {
 
     /// The delegate for (de)multiplexing inbound streams.
     private var inboundStreamMultiplexerState: InboundStreamMultiplexerState
+
+    @usableFromInline
     internal var inboundStreamMultiplexer: InboundStreamMultiplexer? {
         return self.inboundStreamMultiplexerState.multiplexer
     }
@@ -178,7 +180,7 @@ public final class NIOHTTP2Handler: ChannelDuplexHandler {
     }
 
     /// The mode for this parser to operate in: client or server.
-    public enum ParserMode {
+    public enum ParserMode: Sendable {
         /// Client mode
         case client
 
@@ -212,7 +214,9 @@ public final class NIOHTTP2Handler: ChannelDuplexHandler {
                   headerBlockValidation: headerBlockValidation,
                   contentLengthValidation: contentLengthValidation,
                   maximumSequentialEmptyDataFrames: 1,
-                  maximumBufferedControlFrames: 10000)
+                  maximumBufferedControlFrames: 10000,
+                  maximumResetFrameCount: 200,
+                  resetFrameCounterWindow: .seconds(30))
     }
 
     /// Constructs a ``NIOHTTP2Handler``.
@@ -239,8 +243,30 @@ public final class NIOHTTP2Handler: ChannelDuplexHandler {
                   headerBlockValidation: headerBlockValidation,
                   contentLengthValidation: contentLengthValidation,
                   maximumSequentialEmptyDataFrames: maximumSequentialEmptyDataFrames,
-                  maximumBufferedControlFrames: maximumBufferedControlFrames)
+                  maximumBufferedControlFrames: maximumBufferedControlFrames,
+                  maximumResetFrameCount: 200,
+                  resetFrameCounterWindow: .seconds(30))
 
+    }
+
+    /// Constructs a ``NIOHTTP2Handler``.
+    /// 
+    /// - Parameters:
+    ///   - mode: The mode for this handler, client or server.
+    ///   - connectionConfiguration: The settings that will be used when establishing the connection.
+    ///   - streamConfiguration: The settings that will be used when establishing new streams.
+    public convenience init(mode: ParserMode,
+                            connectionConfiguration: ConnectionConfiguration = .init(),
+                            streamConfiguration: StreamConfiguration = .init()) {
+        self.init(mode: mode,
+                  eventLoop: nil,
+                  initialSettings: connectionConfiguration.initialSettings,
+                  headerBlockValidation: connectionConfiguration.headerBlockValidation,
+                  contentLengthValidation: connectionConfiguration.contentLengthValidation,
+                  maximumSequentialEmptyDataFrames: connectionConfiguration.maximumSequentialEmptyDataFrames,
+                  maximumBufferedControlFrames: connectionConfiguration.maximumBufferedControlFrames,
+                  maximumResetFrameCount: streamConfiguration.streamResetFrameRateLimit.maximumCount,
+                  resetFrameCounterWindow: streamConfiguration.streamResetFrameRateLimit.windowLength)
     }
 
     private init(mode: ParserMode,
@@ -249,13 +275,15 @@ public final class NIOHTTP2Handler: ChannelDuplexHandler {
                  headerBlockValidation: ValidationState,
                  contentLengthValidation: ValidationState,
                  maximumSequentialEmptyDataFrames: Int,
-                 maximumBufferedControlFrames: Int) {
-        self.eventLoop = eventLoop
+                 maximumBufferedControlFrames: Int,
+                 maximumResetFrameCount: Int,
+                 resetFrameCounterWindow: TimeAmount) {
+        self._eventLoop = eventLoop
         self.stateMachine = HTTP2ConnectionStateMachine(role: .init(mode), headerBlockValidation: .init(headerBlockValidation), contentLengthValidation: .init(contentLengthValidation))
         self.mode = mode
         self.initialSettings = initialSettings
         self.outboundBuffer = CompoundOutboundBuffer(mode: mode, initialMaxOutboundStreams: 100, maxBufferedControlFrames: maximumBufferedControlFrames)
-        self.denialOfServiceValidator = DOSHeuristics(maximumSequentialEmptyDataFrames: maximumSequentialEmptyDataFrames)
+        self.denialOfServiceValidator = DOSHeuristics(maximumSequentialEmptyDataFrames: maximumSequentialEmptyDataFrames, maximumResetFrameCount: maximumResetFrameCount, resetFrameCounterWindow: resetFrameCounterWindow)
         self.tolerateImpossibleStateTransitionsInDebugMode = false
         self.inboundStreamMultiplexerState = .uninitializedLegacy
     }
@@ -274,19 +302,25 @@ public final class NIOHTTP2Handler: ChannelDuplexHandler {
     ///         upper limit on the depth of this queue. Defaults to 10,000.
     ///   - tolerateImpossibleStateTransitionsInDebugMode: Whether impossible state transitions should be tolerated
     ///         in debug mode.
+    ///   - maximumResetFrameCount: Controls the maximum permitted reset frames within a given time window. Too many may exhaust CPU resources. To protect
+    ///         against this DoS vector we put an upper limit on this rate. Defaults to 200.
+    ///   - resetFrameCounterWindow:  Controls the sliding window used to enforce the maximum permitted reset frames rate. Too many may exhaust CPU resources. To protect
+    ///         against this DoS vector we put an upper limit on this rate. 30 seconds.
     internal init(mode: ParserMode,
                   initialSettings: HTTP2Settings = nioDefaultSettings,
                   headerBlockValidation: ValidationState = .enabled,
                   contentLengthValidation: ValidationState = .enabled,
                   maximumSequentialEmptyDataFrames: Int = 1,
                   maximumBufferedControlFrames: Int = 10000,
-                  tolerateImpossibleStateTransitionsInDebugMode: Bool = false) {
+                  tolerateImpossibleStateTransitionsInDebugMode: Bool = false,
+                  maximumResetFrameCount: Int = 200,
+                  resetFrameCounterWindow: TimeAmount = .seconds(30)) {
         self.stateMachine = HTTP2ConnectionStateMachine(role: .init(mode), headerBlockValidation: .init(headerBlockValidation), contentLengthValidation: .init(contentLengthValidation))
         self.mode = mode
-        self.eventLoop = nil
+        self._eventLoop = nil
         self.initialSettings = initialSettings
         self.outboundBuffer = CompoundOutboundBuffer(mode: mode, initialMaxOutboundStreams: 100, maxBufferedControlFrames: maximumBufferedControlFrames)
-        self.denialOfServiceValidator = DOSHeuristics(maximumSequentialEmptyDataFrames: maximumSequentialEmptyDataFrames)
+        self.denialOfServiceValidator = DOSHeuristics(maximumSequentialEmptyDataFrames: maximumSequentialEmptyDataFrames, maximumResetFrameCount: maximumResetFrameCount, resetFrameCounterWindow: resetFrameCounterWindow)
         self.tolerateImpossibleStateTransitionsInDebugMode = tolerateImpossibleStateTransitionsInDebugMode
         self.inboundStreamMultiplexerState = .uninitializedLegacy
     }
@@ -854,9 +888,9 @@ extension NIOHTTP2Handler {
             // Don't allow infinite recursion through this method.
             return
         }
-        
+
         self.isUnbufferingAndFlushingAutomaticFrames = true
-    
+
         loop: while true {
             switch self.outboundBuffer.nextFlushedWritableFrame(channelWritable: self.channelWritable) {
             case .noFrame:
@@ -867,7 +901,7 @@ extension NIOHTTP2Handler {
                 self.processOutboundFrame(context: context, frame: frame, promise: promise)
             }
         }
-        
+
         self.isUnbufferingAndFlushingAutomaticFrames = false
         self.flushIfNecessary(context: context)
     }
@@ -1004,17 +1038,11 @@ extension NIOHTTP2Handler {
 
 
 extension NIOHTTP2Handler {
-#if swift(>=5.7)
     /// The type of all `inboundStreamInitializer` callbacks which do not need to return data.
     public typealias StreamInitializer = NIOChannelInitializer
     /// The type of NIO Channel initializer callbacks which need to return untyped data.
+    @usableFromInline
     internal typealias StreamInitializerWithAnyOutput = @Sendable (Channel) -> EventLoopFuture<any Sendable>
-#else
-    /// The type of all `inboundStreamInitializer` callbacks which need to return data.
-    public typealias StreamInitializer = NIOChannelInitializer
-    /// The type of NIO Channel initializer callbacks which need to return untyped data.
-    internal typealias StreamInitializerWithAnyOutput = (Channel) -> EventLoopFuture<any Sendable>
-#endif
 
     /// Creates a new ``NIOHTTP2Handler`` with a local multiplexer. (i.e. using
     /// ``StreamMultiplexer``.)
@@ -1043,12 +1071,15 @@ extension NIOHTTP2Handler {
             headerBlockValidation: connectionConfiguration.headerBlockValidation,
             contentLengthValidation: connectionConfiguration.contentLengthValidation,
             maximumSequentialEmptyDataFrames: connectionConfiguration.maximumSequentialEmptyDataFrames,
-            maximumBufferedControlFrames: connectionConfiguration.maximumBufferedControlFrames
+            maximumBufferedControlFrames: connectionConfiguration.maximumBufferedControlFrames,
+            maximumResetFrameCount: streamConfiguration.streamResetFrameRateLimit.maximumCount,
+            resetFrameCounterWindow: streamConfiguration.streamResetFrameRateLimit.windowLength
         )
 
         self.inboundStreamMultiplexerState = .uninitializedInline(streamConfiguration, inboundStreamInitializer, streamDelegate)
     }
 
+    @usableFromInline
     internal convenience init(
         mode: ParserMode,
         eventLoop: EventLoop,
@@ -1064,7 +1095,9 @@ extension NIOHTTP2Handler {
             headerBlockValidation: connectionConfiguration.headerBlockValidation,
             contentLengthValidation: connectionConfiguration.contentLengthValidation,
             maximumSequentialEmptyDataFrames: connectionConfiguration.maximumSequentialEmptyDataFrames,
-            maximumBufferedControlFrames: connectionConfiguration.maximumBufferedControlFrames
+            maximumBufferedControlFrames: connectionConfiguration.maximumBufferedControlFrames,
+            maximumResetFrameCount: streamConfiguration.streamResetFrameRateLimit.maximumCount,
+            resetFrameCounterWindow: streamConfiguration.streamResetFrameRateLimit.windowLength
         )
         self.inboundStreamMultiplexerState = .uninitializedAsync(streamConfiguration, inboundStreamInitializerWithAnyOutput, streamDelegate)
     }
@@ -1089,6 +1122,17 @@ extension NIOHTTP2Handler {
         public var targetWindowSize: Int = 65535
         public var outboundBufferSizeHighWatermark: Int = 8196
         public var outboundBufferSizeLowWatermark: Int = 4092
+        public var streamResetFrameRateLimit: StreamResetFrameRateLimitConfiguration = .init()
+        public init() {}
+    }
+    
+    /// Stream reset frame rate limit configuration.
+    ///
+    /// The settings that control the maximum permitted reset frames within a given time window. Too many may exhaust CPU resources.
+    /// To protect against this DoS vector we put an upper limit on this rate.
+    public struct StreamResetFrameRateLimitConfiguration: Hashable, Sendable {
+        public var maximumCount: Int = 200
+        public var windowLength: TimeAmount = .seconds(30)
         public init() {}
     }
 
@@ -1117,13 +1161,14 @@ extension NIOHTTP2Handler {
     /// i.e. it was initialized with an `inboundStreamInitializer`.
     public var multiplexer: EventLoopFuture<StreamMultiplexer> {
         // We need to return a future here so that we can synchronize access on the underlying `self.inboundStreamMultiplexer`
-        if self.eventLoop!.inEventLoop {
-            return self.eventLoop!.makeCompletedFuture {
+        if self._eventLoop!.inEventLoop {
+            return self._eventLoop!.makeCompletedFuture {
                 return try self.syncMultiplexer()
             }
         } else {
-            return self.eventLoop!.submit {
-                return try self.syncMultiplexer()
+            let unsafeSelf = UnsafeTransfer(self)
+            return self._eventLoop!.submit {
+                return try unsafeSelf.wrappedValue.syncMultiplexer()
             }
         }
     }
@@ -1134,24 +1179,31 @@ extension NIOHTTP2Handler {
     /// > - The ``NIOHTTP2Handler`` uses a local multiplexer, i.e. it was initialized with an `inboundStreamInitializer`.
     /// > - The caller is already on the correct event loop.
     public func syncMultiplexer() throws -> StreamMultiplexer {
-        self.eventLoop!.preconditionInEventLoop()
+        self._eventLoop!.preconditionInEventLoop()
+        guard let inboundStreamMultiplexer = self.inboundStreamMultiplexer else {
+            throw NIOHTTP2Errors.missingMultiplexer()
+        }
 
-        switch self.inboundStreamMultiplexer {
-        case let .some(.inline(multiplexer)):
-            return StreamMultiplexer(multiplexer)
-        case .some(.legacy), .none:
+        switch inboundStreamMultiplexer {
+        case let .inline(multiplexer):
+            return StreamMultiplexer(multiplexer, eventLoop: self._eventLoop!)
+        case .legacy:
             throw NIOHTTP2Errors.missingMultiplexer()
         }
     }
 
+    @inlinable
     @available(macOS 10.15, iOS 13.0, watchOS 6.0, tvOS 13.0, *)
-    internal func syncAsyncStreamMultiplexer<Output: Sendable>(continuation: any AnyContinuation, inboundStreamChannels: NIOHTTP2InboundStreamChannels<Output>) throws -> AsyncStreamMultiplexer<Output> {
-        self.eventLoop!.preconditionInEventLoop()
+    internal func syncAsyncStreamMultiplexer<Output: Sendable>(continuation: any AnyContinuation, inboundStreamChannels: NIOHTTP2AsyncSequence<Output>) throws -> AsyncStreamMultiplexer<Output> {
+        self._eventLoop!.preconditionInEventLoop()
+        guard let inboundStreamMultiplexer = self.inboundStreamMultiplexer else {
+            throw NIOHTTP2Errors.missingMultiplexer()
+        }
 
-        switch self.inboundStreamMultiplexer {
-        case let .some(.inline(multiplexer)):
-            return AsyncStreamMultiplexer(multiplexer, continuation: continuation, inboundStreamChannels: inboundStreamChannels)
-        case .some(.legacy), .none:
+        switch inboundStreamMultiplexer {
+        case let .inline(multiplexer):
+            return AsyncStreamMultiplexer(multiplexer, eventLoop: self._eventLoop!, continuation: continuation, inboundStreamChannels: inboundStreamChannels)
+        case .legacy:
             throw NIOHTTP2Errors.missingMultiplexer()
         }
     }
