@@ -15,13 +15,14 @@
 import XCTest
 
 import NIOConcurrencyHelpers
-@_spi(AsyncChannel) import NIOCore
+import NIOCore
 import NIOEmbedded
 import NIOHPACK
 import NIOHTTP1
-@_spi(AsyncChannel) import NIOHTTP2
+import NIOHTTP2
 import NIOTLS
 
+@available(macOS 10.15, iOS 13.0, watchOS 6.0, tvOS 13.0, *)
 final class ConfiguringPipelineAsyncMultiplexerTests: XCTestCase {
     var clientChannel: NIOAsyncTestingChannel!
     var serverChannel: NIOAsyncTestingChannel!
@@ -91,6 +92,70 @@ final class ConfiguringPipelineAsyncMultiplexerTests: XCTestCase {
         }
     }
 
+    func testCancellingAsyncStreamConsumer() async throws {
+        let requestCount = 200
+
+        let serverRecorder = InboundFramePayloadRecorder()
+
+        let clientMultiplexer = try await assertNoThrowWithValue(
+            try await self.clientChannel.configureAsyncHTTP2Pipeline(mode: .client) { channel -> EventLoopFuture<Channel> in
+                channel.eventLoop.makeSucceededFuture(channel)
+            }.get()
+        )
+
+        let serverMultiplexer = try await assertNoThrowWithValue(
+            try await self.serverChannel.configureAsyncHTTP2Pipeline(mode: .server) { channel -> EventLoopFuture<Channel> in
+                channel.pipeline.addHandlers([OKResponder(), serverRecorder]).map { _ in channel }
+            }.get()
+        )
+
+        try await assertNoThrow(try await self.assertDoHandshake(client: self.clientChannel, server: self.serverChannel))
+
+        // Launch a server
+        let serverTask = Task {
+            var serverInboundChannelCount = 0
+            for try await _ in serverMultiplexer.inbound {
+                serverInboundChannelCount += 1
+            }
+
+            try Task.checkCancellation()
+
+            return serverInboundChannelCount
+        }
+
+        // client
+        for i in 0 ..< requestCount {
+            // Let's try sending some requests.
+            let streamChannel = try await clientMultiplexer.openStream { channel -> EventLoopFuture<Channel> in
+                return channel.pipeline.addHandlers([SimpleRequest(), InboundFramePayloadRecorder()]).map {
+                    return channel
+                }
+            }
+
+            // When we get above 100, cancel the server task.
+            if i == 100 { serverTask.cancel() }
+
+            let clientRecorder = try await streamChannel.pipeline.handler(type: InboundFramePayloadRecorder.self).get()
+            try await Self.deliverAllBytes(from: self.clientChannel, to: self.serverChannel)
+            try await Self.deliverAllBytes(from: self.serverChannel, to: self.clientChannel)
+
+            clientRecorder.receivedFrames.assertFramePayloadsMatch([ConfiguringPipelineAsyncMultiplexerTests.responseFramePayload])
+            try await streamChannel.closeFuture.get()
+        }
+
+        try await assertNoThrow(try await self.clientChannel.finish())
+        try await assertNoThrow(try await self.serverChannel.finish())
+
+        do {
+            _ = try await serverTask.value
+            XCTFail("Server unexpectedly succeeded")
+        } catch is CancellationError {
+            // Expected
+        } catch {
+            XCTFail("Unexpected error throw: \(error)")
+        }
+    }
+
     // `testBasicPipelineCommunicates` ensures that a client-server system set up to use async stream abstractions
     // can communicate successfully.
     func testBasicPipelineCommunicates() async throws {
@@ -125,15 +190,15 @@ final class ConfiguringPipelineAsyncMultiplexerTests: XCTestCase {
             // client
             for _ in 0 ..< requestCount {
                 // Let's try sending some requests
-                let streamChannel = try await clientMultiplexer.createStreamChannel { channel -> EventLoopFuture<Channel> in
+                let streamChannel = try await clientMultiplexer.openStream { channel -> EventLoopFuture<Channel> in
                     return channel.pipeline.addHandlers([SimpleRequest(), InboundFramePayloadRecorder()]).map {
                         return channel
                     }
                 }
 
                 let clientRecorder = try await streamChannel.pipeline.handler(type: InboundFramePayloadRecorder.self).get()
-                try await self.deliverAllBytes(from: self.clientChannel, to: self.serverChannel)
-                try await self.deliverAllBytes(from: self.serverChannel, to: self.clientChannel)
+                try await Self.deliverAllBytes(from: self.clientChannel, to: self.serverChannel)
+                try await Self.deliverAllBytes(from: self.serverChannel, to: self.clientChannel)
                 clientRecorder.receivedFrames.assertFramePayloadsMatch([ConfiguringPipelineAsyncMultiplexerTests.responseFramePayload])
                 try await streamChannel.closeFuture.get()
             }
@@ -150,13 +215,14 @@ final class ConfiguringPipelineAsyncMultiplexerTests: XCTestCase {
 
     // `testNIOAsyncConnectionStreamChannelPipelineCommunicates` ensures that a client-server system set up to use `NIOAsyncChannel`
     // wrappers around connection and stream channels can communicate successfully.
+    @available(*, deprecated, message: "Deprecated so deprecated functionality can be tested without warnings")
     func testNIOAsyncConnectionStreamChannelPipelineCommunicates() async throws {
         let requestCount = 100
 
         let clientMultiplexer = try await assertNoThrowWithValue(
             try await self.clientChannel.configureAsyncHTTP2Pipeline(
                 mode: .client,
-                inboundStreamInitializer: { channel in
+                streamInitializer: { channel in
                     channel.eventLoop.makeCompletedFuture {
                         try NIOAsyncChannel(
                             synchronouslyWrapping: channel,
@@ -167,34 +233,35 @@ final class ConfiguringPipelineAsyncMultiplexerTests: XCTestCase {
             ).get()
         )
 
-        let serverMultiplexer = try await assertNoThrowWithValue(
-            try await self.serverChannel.configureAsyncHTTP2Pipeline(
-                mode: .server,
-                inboundStreamInitializer: { channel in
-                    channel.eventLoop.makeCompletedFuture {
-                        try NIOAsyncChannel(
-                            synchronouslyWrapping: channel,
-                            configuration: .init(inboundType: HTTP2Frame.FramePayload.self, outboundType: HTTP2Frame.FramePayload.self)
-                        )
-                    }
+        let serverMultiplexerFuture = self.serverChannel.configureAsyncHTTP2Pipeline(
+            mode: .server,
+            streamInitializer: { channel in
+                channel.eventLoop.makeCompletedFuture {
+                    try NIOAsyncChannel(
+                        synchronouslyWrapping: channel,
+                        configuration: .init(inboundType: HTTP2Frame.FramePayload.self, outboundType: HTTP2Frame.FramePayload.self)
+                    )
                 }
-            ).get()
+            }
         )
 
         try await assertNoThrow(try await self.assertDoHandshake(client: self.clientChannel, server: self.serverChannel))
 
+        let serverChannel = self.serverChannel!
+        let clientChannel = self.clientChannel!
         try await withThrowingTaskGroup(of: Int.self, returning: Void.self) { group in
             // server
             group.addTask {
+                let serverMultiplexer = try await assertNoThrowWithValue(try await serverMultiplexerFuture.get())
                 var serverInboundChannelCount = 0
                 for try await streamChannel in serverMultiplexer.inbound {
-                    for try await receivedFrame in streamChannel.inboundStream {
+                    for try await receivedFrame in streamChannel.inbound {
                         receivedFrame.assertFramePayloadMatches(this: ConfiguringPipelineAsyncMultiplexerTests.requestFramePayload)
 
-                        try await streamChannel.outboundWriter.write(ConfiguringPipelineAsyncMultiplexerTests.responseFramePayload)
-                        streamChannel.outboundWriter.finish()
+                        try await streamChannel.outbound.write(ConfiguringPipelineAsyncMultiplexerTests.responseFramePayload)
+                        streamChannel.outbound.finish()
 
-                        try await self.deliverAllBytes(from: self.serverChannel, to: self.clientChannel)
+                        try await Self.deliverAllBytes(from: serverChannel, to: clientChannel)
                     }
                     serverInboundChannelCount += 1
                 }
@@ -203,7 +270,7 @@ final class ConfiguringPipelineAsyncMultiplexerTests: XCTestCase {
 
             // client
             for _ in 0 ..< requestCount {
-                let streamChannel = try await clientMultiplexer.createStreamChannel() { channel in
+                let streamChannel = try await clientMultiplexer.openStream() { channel in
                     channel.eventLoop.makeCompletedFuture {
                         try NIOAsyncChannel(
                             synchronouslyWrapping: channel,
@@ -215,12 +282,12 @@ final class ConfiguringPipelineAsyncMultiplexerTests: XCTestCase {
                     }
                 }
                 // Let's try sending some requests
-                try await streamChannel.outboundWriter.write(ConfiguringPipelineAsyncMultiplexerTests.requestFramePayload)
-                streamChannel.outboundWriter.finish()
+                try await streamChannel.outbound.write(ConfiguringPipelineAsyncMultiplexerTests.requestFramePayload)
+                streamChannel.outbound.finish()
 
-                try await self.deliverAllBytes(from: self.clientChannel, to: self.serverChannel)
+                try await Self.deliverAllBytes(from: self.clientChannel, to: self.serverChannel)
 
-                for try await receivedFrame in streamChannel.inboundStream {
+                for try await receivedFrame in streamChannel.inbound {
                     receivedFrame.assertFramePayloadMatches(this: ConfiguringPipelineAsyncMultiplexerTests.responseFramePayload)
                 }
             }
@@ -232,7 +299,7 @@ final class ConfiguringPipelineAsyncMultiplexerTests: XCTestCase {
             XCTAssertEqual(serverInboundChannelCount, requestCount, "We should have created one server-side channel as a result of the one HTTP/2 stream used.")
         }
     }
-    
+
     // `testNegotiatedHTTP2BasicPipelineCommunicates` ensures that a client-server system set up to use async stream abstractions
     // can communicate successfully when HTTP/2 is negotiated.
     func testNegotiatedHTTP2BasicPipelineCommunicates() async throws {
@@ -246,32 +313,31 @@ final class ConfiguringPipelineAsyncMultiplexerTests: XCTestCase {
             }.get()
         )
 
-        let nioProtocolNegotiationResult = try await self.serverChannel.configureAsyncHTTPServerPipeline() { channel in
+        let negotiationResultFuture = try await self.serverChannel.configureAsyncHTTPServerPipeline() { channel in
             channel.eventLoop.makeSucceededVoidFuture()
         } http2ConnectionInitializer: { channel in
             channel.eventLoop.makeSucceededVoidFuture()
-        } http2InboundStreamInitializer: { channel -> EventLoopFuture<Channel> in
+        } http2StreamInitializer: { channel -> EventLoopFuture<Channel> in
             channel.pipeline.addHandlers([OKResponder(), serverRecorder]).map { _ in channel }
         }.get()
 
         // Let's pretend the TLS handler did protocol negotiation for us
         self.serverChannel.pipeline.fireUserInboundEventTriggered(TLSUserEvent.handshakeCompleted(negotiatedProtocol: "h2"))
 
-        let negotiationResult = try await nioProtocolNegotiationResult.getResult()
-
         try await assertNoThrow(try await self.assertDoHandshake(client: self.clientChannel, server: self.serverChannel))
-
-        let serverMultiplexer: NIOHTTP2Handler.AsyncStreamMultiplexer<Channel>
-        switch negotiationResult {
-        case .http1_1:
-            preconditionFailure("Negotiation result must be HTTP/2")
-        case .http2(let (_, multiplexer)):
-            serverMultiplexer = multiplexer
-        }
 
         try await withThrowingTaskGroup(of: Int.self, returning: Void.self) { group in
             // server
             group.addTask {
+                let negotiationResult = try await negotiationResultFuture.get()
+                let serverMultiplexer: NIOHTTP2Handler.AsyncStreamMultiplexer<Channel>
+                switch negotiationResult {
+                case .http1_1:
+                    preconditionFailure("Negotiation result must be HTTP/2")
+                case .http2(let (_, multiplexer)):
+                    serverMultiplexer = multiplexer
+                }
+
                 var serverInboundChannelCount = 0
                 for try await _ in serverMultiplexer.inbound {
                     serverInboundChannelCount += 1
@@ -282,7 +348,7 @@ final class ConfiguringPipelineAsyncMultiplexerTests: XCTestCase {
             // client
             for _ in 0 ..< requestCount {
                 // Let's try sending some requests
-                let streamChannel = try await clientMultiplexer.createStreamChannel { channel -> EventLoopFuture<Channel> in
+                let streamChannel = try await clientMultiplexer.openStream { channel -> EventLoopFuture<Channel> in
                     return channel.pipeline.addHandlers([SimpleRequest(), InboundFramePayloadRecorder()]).map {
                         return channel
                     }
@@ -290,8 +356,8 @@ final class ConfiguringPipelineAsyncMultiplexerTests: XCTestCase {
 
                 let clientRecorder = try await streamChannel.pipeline.handler(type: InboundFramePayloadRecorder.self).get()
 
-                try await self.deliverAllBytes(from: self.clientChannel, to: self.serverChannel)
-                try await self.deliverAllBytes(from: self.serverChannel, to: self.clientChannel)
+                try await Self.deliverAllBytes(from: self.clientChannel, to: self.serverChannel)
+                try await Self.deliverAllBytes(from: self.serverChannel, to: self.clientChannel)
 
                 clientRecorder.receivedFrames.assertFramePayloadsMatch([ConfiguringPipelineAsyncMultiplexerTests.responseFramePayload])
                 try await streamChannel.closeFuture.get()
@@ -312,25 +378,25 @@ final class ConfiguringPipelineAsyncMultiplexerTests: XCTestCase {
     func testNegotiatedHTTP1BasicPipelineCommunicates() async throws {
         let requestCount = 100
 
-        let _ = try await self.clientChannel.pipeline.addHTTPClientHandlers().map { _ in
-            self.clientChannel.pipeline.addHandlers([InboundRecorderHandler<HTTPClientResponsePart>(), HTTP1ClientSendability()])
+        let _ = try await self.clientChannel.pipeline.addHTTPClientHandlers().map { [clientChannel = self.clientChannel!] _ in
+            clientChannel.pipeline.addHandlers([InboundRecorderHandler<HTTPClientResponsePart>(), HTTP1ClientSendability()])
         }.get()
 
-        let nioProtocolNegotiationResult = try await self.serverChannel.configureAsyncHTTPServerPipeline() { channel in
+        let negotiationResultFuture = try await self.serverChannel.configureAsyncHTTPServerPipeline() { channel in
             channel.pipeline.addHandlers([HTTP1OKResponder(), InboundRecorderHandler<HTTPServerRequestPart>()])
         } http2ConnectionInitializer: { channel in
             channel.eventLoop.makeSucceededVoidFuture()
-        } http2InboundStreamInitializer: { channel -> EventLoopFuture<Channel> in
+        } http2StreamInitializer: { channel -> EventLoopFuture<Channel> in
             channel.eventLoop.makeSucceededFuture(channel)
         }.get()
 
         // Let's pretend the TLS handler did protocol negotiation for us
         self.serverChannel.pipeline.fireUserInboundEventTriggered(TLSUserEvent.handshakeCompleted(negotiatedProtocol: "http/1.1"))
 
-        let negotiationResult = try await nioProtocolNegotiationResult.getResult()
+        let negotiationResult = try await negotiationResultFuture.get()
 
-        try await self.deliverAllBytes(from: self.clientChannel, to: self.serverChannel)
-        try await self.deliverAllBytes(from: self.serverChannel, to: self.clientChannel)
+        try await Self.deliverAllBytes(from: self.clientChannel, to: self.serverChannel)
+        try await Self.deliverAllBytes(from: self.serverChannel, to: self.clientChannel)
 
         switch negotiationResult {
         case .http1_1:
@@ -345,23 +411,31 @@ final class ConfiguringPipelineAsyncMultiplexerTests: XCTestCase {
             // we need to put these through a mapping to remove references to `IOData` which isn't Sendable
             try await self.clientChannel.writeOutbound(HTTP1ClientSendability.RequestPart.head(ConfiguringPipelineAsyncMultiplexerTests.requestHead))
             try await self.clientChannel.writeOutbound(HTTP1ClientSendability.RequestPart.end(nil))
-            try await self.deliverAllBytes(from: self.clientChannel, to: self.serverChannel)
-            try await self.deliverAllBytes(from: self.serverChannel, to: self.clientChannel)
+            try await Self.deliverAllBytes(from: self.clientChannel, to: self.serverChannel)
+            try await Self.deliverAllBytes(from: self.serverChannel, to: self.clientChannel)
         }
 
         // check expectations
         let clientRecorder = try await self.clientChannel.pipeline.handler(type: InboundRecorderHandler<HTTPClientResponsePart>.self).get()
         let serverRecorder = try await self.serverChannel.pipeline.handler(type: InboundRecorderHandler<HTTPServerRequestPart>.self).get()
 
-        XCTAssertEqual(serverRecorder.receivedParts.count, requestCount*2)
-        XCTAssertEqual(clientRecorder.receivedParts.count, requestCount*2)
+        serverRecorder.receivedParts.withLockedValue { receivedParts in
+            XCTAssertEqual(receivedParts.count, requestCount*2)
+        }
+        clientRecorder.receivedParts.withLockedValue { receivedParts in
+            XCTAssertEqual(receivedParts.count, requestCount*2)
+        }
 
         for i in 0 ..< requestCount {
-            XCTAssertEqual(serverRecorder.receivedParts[i*2], HTTPServerRequestPart.head(ConfiguringPipelineAsyncMultiplexerTests.requestHead), "Unexpected request part in iteration \(i)")
-            XCTAssertEqual(serverRecorder.receivedParts[i*2+1], HTTPServerRequestPart.end(nil), "Unexpected request part in iteration \(i)")
+            serverRecorder.receivedParts.withLockedValue { receivedParts in
+                XCTAssertEqual(receivedParts[i*2], HTTPServerRequestPart.head(ConfiguringPipelineAsyncMultiplexerTests.requestHead), "Unexpected request part in iteration \(i)")
+                XCTAssertEqual(receivedParts[i*2+1], HTTPServerRequestPart.end(nil), "Unexpected request part in iteration \(i)")
+            }
 
-            XCTAssertEqual(clientRecorder.receivedParts[i*2], HTTPClientResponsePart.head(ConfiguringPipelineAsyncMultiplexerTests.responseHead), "Unexpected response part in iteration \(i)")
-            XCTAssertEqual(clientRecorder.receivedParts[i*2+1], HTTPClientResponsePart.end(nil), "Unexpected response part in iteration \(i)")
+            clientRecorder.receivedParts.withLockedValue { receivedParts in
+                XCTAssertEqual(receivedParts[i*2], HTTPClientResponsePart.head(ConfiguringPipelineAsyncMultiplexerTests.responseHead), "Unexpected response part in iteration \(i)")
+                XCTAssertEqual(receivedParts[i*2+1], HTTPClientResponsePart.end(nil), "Unexpected response part in iteration \(i)")
+            }
         }
 
         try await assertNoThrow(try await self.clientChannel.finish())
@@ -417,27 +491,15 @@ final class ConfiguringPipelineAsyncMultiplexerTests: XCTestCase {
     }
 
     /// A simple channel handler that records inbound messages.
-    internal final class InboundRecorderHandler<message>: ChannelInboundHandler, @unchecked Sendable {
+    internal final class InboundRecorderHandler<message: Sendable>: ChannelInboundHandler, Sendable {
         typealias InboundIn = message
 
-        private let partsLock = NIOLock()
-        private var _receivedParts: [message] = []
-
-        var receivedParts: [message] {
-            get {
-                self.partsLock.withLock {
-                    self._receivedParts
-                }
-            }
-            set {
-                self.partsLock.withLock {
-                    self._receivedParts = newValue
-                }
-            }
-        }
+        let receivedParts = NIOLockedValueBox<[message]>([])
 
         func channelRead(context: ChannelHandlerContext, data: NIOAny) {
-            self.receivedParts.append(self.unwrapInboundIn(data))
+            self.receivedParts.withLockedValue { receivedParts in
+                receivedParts.append(self.unwrapInboundIn(data))
+            }
             context.fireChannelRead(data)
         }
     }
@@ -446,7 +508,7 @@ final class ConfiguringPipelineAsyncMultiplexerTests: XCTestCase {
 #if swift(<5.9)
 // this should be available in the std lib from 5.9 onwards
 extension AsyncStream {
-    fileprivate static func makeStream(
+    internal static func makeStream(
         of elementType: Element.Type = Element.self,
         bufferingPolicy limit: Continuation.BufferingPolicy = .unbounded
     ) -> (stream: AsyncStream<Element>, continuation: AsyncStream<Element>.Continuation) {
