@@ -13,6 +13,7 @@
 //===----------------------------------------------------------------------===//
 
 import XCTest
+import Atomics
 import NIOConcurrencyHelpers
 import NIOCore
 import NIOEmbedded
@@ -119,14 +120,16 @@ typealias InboundFrameRecorder = InboundRecorder<HTTP2Frame>
 typealias InboundFramePayloadRecorder = InboundRecorder<HTTP2Frame.FramePayload>
 
 /// A handler that tracks the number of times read() was called on the channel.
-final class ReadCounter: ChannelOutboundHandler {
+final class ReadCounter: ChannelOutboundHandler, Sendable {
     typealias OutboundIn = Any
     typealias OutboundOut = Any
 
-    var readCount = 0
+    let readCount = NIOLockedValueBox<Int>(0)
 
     func read(context: ChannelHandlerContext) {
-        readCount += 1
+        self.readCount.withLockedValue { readCount in
+            readCount += 1
+        }
         context.read()
     }
 }
@@ -146,14 +149,16 @@ final class FlushCounter: ChannelOutboundHandler {
 }
 
 
-final class ReadCompleteCounter: ChannelInboundHandler {
+final class ReadCompleteCounter: ChannelInboundHandler, Sendable {
     typealias InboundIn = Any
     typealias InboundOut = Any
 
-    var readCompleteCount = 0
+    let readCompleteCount = NIOLockedValueBox<Int>(0)
 
     func channelReadComplete(context: ChannelHandlerContext) {
-        self.readCompleteCount += 1
+        self.readCompleteCount.withLockedValue { readCompleteCount in
+            readCompleteCount += 1
+        }
         context.fireChannelReadComplete()
     }
 }
@@ -266,9 +271,9 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
 
     @available(*, deprecated, message: "Deprecated so deprecated functionality can be tested without warnings")
     func testHeadersFramesCreateNewChannels() throws {
-        var channelCount = 0
+        let channelCount = ManagedAtomic<Int>(0)
         let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { (channel, _) in
-            channelCount += 1
+            channelCount.wrappingIncrement(ordering: .sequentiallyConsistent)
             return channel.close()
         }
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(multiplexer).wait())
@@ -279,17 +284,15 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
             XCTAssertNoThrow(try self.channel.writeInbound(frame))
         }
 
-        XCTAssertEqual(channelCount, 50)
+        XCTAssertEqual(channelCount.load(ordering: .sequentiallyConsistent), 50)
         XCTAssertNoThrow(try self.channel.finish())
     }
 
     @available(*, deprecated, message: "Deprecated so deprecated functionality can be tested without warnings")
     func testChannelsCloseThemselvesWhenToldTo() throws {
-        var completedChannelCount = 0
-        var closeFutures: [EventLoopFuture<Void>] = []
+        let completedChannelCount = ManagedAtomic<Int>(0)
         let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { (channel, _) in
-            closeFutures.append(channel.closeFuture)
-            channel.closeFuture.whenSuccess { completedChannelCount += 1 }
+            channel.closeFuture.whenSuccess { completedChannelCount.wrappingIncrement(ordering: .sequentiallyConsistent) }
             return channel.eventLoop.makeSucceededFuture(())
         }
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(multiplexer).wait())
@@ -300,7 +303,7 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
             let frame = HTTP2Frame(streamID: streamID, payload: .headers(.init(headers: HPACKHeaders(), endStream: true)))
             XCTAssertNoThrow(try self.channel.writeInbound(frame))
         }
-        XCTAssertEqual(completedChannelCount, 0)
+        XCTAssertEqual(completedChannelCount.load(ordering: .sequentiallyConsistent), 0)
 
         // Now we send them all a clean exit.
         for streamID in streamIDs {
@@ -310,13 +313,13 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         (self.channel.eventLoop as! EmbeddedEventLoop).run()
 
         // At this stage all the promises should be completed.
-        XCTAssertEqual(completedChannelCount, 50)
+        XCTAssertEqual(completedChannelCount.load(ordering: .sequentiallyConsistent), 50)
         XCTAssertNoThrow(try self.channel.finish())
     }
 
     @available(*, deprecated, message: "Deprecated so deprecated functionality can be tested without warnings")
     func testChannelsCloseAfterResetStreamFrameFirstThenEvent() throws {
-        var closeError: Error? = nil
+        let closeError = NIOLockedValueBox<Error?>(nil)
 
         XCTAssertNoThrow(try self.channel.connect(to: SocketAddress(unixDomainSocketPath: "/whatever"), promise: nil))
 
@@ -326,8 +329,14 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         let rstStreamFrame = HTTP2Frame(streamID: streamID, payload: .rstStream(.cancel))
 
         let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { (channel, _) in
-            XCTAssertNil(closeError)
-            channel.closeFuture.whenFailure { closeError = $0 }
+            closeError.withLockedValue { closeError in
+                XCTAssertNil(closeError)
+            }
+            channel.closeFuture.whenFailure { error in
+                closeError.withLockedValue { closeError in
+                    closeError = error
+                }
+            }
             return channel.pipeline.addHandler(FrameExpecter(expectedFrames: [frame, rstStreamFrame]))
         }
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(multiplexer).wait())
@@ -335,7 +344,9 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         // Let's open the stream up.
         XCTAssertNoThrow(try self.channel.writeInbound(frame))
         self.activateStream(streamID)
-        XCTAssertNil(closeError)
+        closeError.withLockedValue { closeError in
+            XCTAssertNil(closeError)
+        }
 
         // Now we can send a RST_STREAM frame.
         XCTAssertNoThrow(try self.channel.writeInbound(rstStreamFrame))
@@ -346,14 +357,16 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         (self.channel.eventLoop as! EmbeddedEventLoop).run()
 
         // At this stage the stream should be closed with the appropriate error code.
-        XCTAssertEqual(closeError as? NIOHTTP2Errors.StreamClosed,
-                       NIOHTTP2Errors.StreamClosed(streamID: streamID, errorCode: .cancel))
+        closeError.withLockedValue { closeError in
+            XCTAssertEqual(closeError as? NIOHTTP2Errors.StreamClosed,
+                           NIOHTTP2Errors.StreamClosed(streamID: streamID, errorCode: .cancel))
+        }
         XCTAssertNoThrow(try self.channel.finish())
     }
 
     @available(*, deprecated, message: "Deprecated so deprecated functionality can be tested without warnings")
     func testChannelsCloseAfterGoawayFrameFirstThenEvent() throws {
-        var closeError: Error? = nil
+        let closeError = NIOLockedValueBox<Error?>(nil)
 
         XCTAssertNoThrow(try self.channel.connect(to: SocketAddress(unixDomainSocketPath: "/whatever"), promise: nil))
 
@@ -363,8 +376,14 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         let goAwayFrame = HTTP2Frame(streamID: .rootStream, payload: .goAway(lastStreamID: .rootStream, errorCode: .http11Required, opaqueData: nil))
 
         let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { (channel, _) in
-            XCTAssertNil(closeError)
-            channel.closeFuture.whenFailure { closeError = $0 }
+            closeError.withLockedValue { closeError in
+                XCTAssertNil(closeError)
+            }
+            channel.closeFuture.whenFailure { error in
+                closeError.withLockedValue { closeError in
+                    closeError = error
+                }
+            }
             // The channel won't see the goaway frame.
             return channel.pipeline.addHandler(FrameExpecter(expectedFrames: [frame]))
         }
@@ -373,7 +392,9 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         // Let's open the stream up.
         XCTAssertNoThrow(try self.channel.writeInbound(frame))
         self.activateStream(streamID)
-        XCTAssertNil(closeError)
+        closeError.withLockedValue { closeError in
+            XCTAssertNil(closeError)
+        }
 
         // Now we can send a GOAWAY frame. This will close the stream.
         XCTAssertNoThrow(try self.channel.writeInbound(goAwayFrame))
@@ -384,8 +405,10 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         (self.channel.eventLoop as! EmbeddedEventLoop).run()
 
         // At this stage the stream should be closed with the appropriate manufactured error code.
-        XCTAssertEqual(closeError as? NIOHTTP2Errors.StreamClosed,
-                       NIOHTTP2Errors.StreamClosed(streamID: streamID, errorCode: .refusedStream))
+        closeError.withLockedValue { closeError in
+            XCTAssertEqual(closeError as? NIOHTTP2Errors.StreamClosed,
+                           NIOHTTP2Errors.StreamClosed(streamID: streamID, errorCode: .refusedStream))
+        }
         XCTAssertNoThrow(try self.channel.finish())
     }
 
@@ -401,7 +424,7 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         XCTAssertThrowsError(try self.channel.writeInbound(dataFrame)) { error in
             XCTAssertEqual(streamID, (error as? NIOHTTP2Errors.NoSuchStream).map { $0.streamID })
         }
-        
+
         self.channel.assertNoFramesReceived()
 
         XCTAssertNoThrow(try self.channel.finish())
@@ -516,16 +539,18 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         XCTAssertTrue(childChannel.isActive)
 
         // Now we close it. This triggers a RST_STREAM frame. The channel will not be closed at this time.
-        var closed = false
-        childChannel.close().whenComplete { _ in closed = true }
+        let closed = ManagedAtomic<Bool>(false)
+        childChannel.close().whenComplete { _ in
+            closed.store(true, ordering: .sequentiallyConsistent)
+        }
         XCTAssertEqual(frameReceiver.flushedWrites.count, 1)
         frameReceiver.flushedWrites[0].assertRstStreamFrame(streamID: streamID, errorCode: .cancel)
-        XCTAssertFalse(closed)
+        XCTAssertFalse(closed.load(ordering: .sequentiallyConsistent))
 
         // Now send the stream closed event. This will satisfy the close promise.
         let userEvent = StreamClosedEvent(streamID: streamID, reason: .cancel)
         self.channel.pipeline.fireUserInboundEventTriggered(userEvent)
-        XCTAssertTrue(closed)
+        XCTAssertTrue(closed.load(ordering: .sequentiallyConsistent))
 
         XCTAssertNoThrow(try self.channel.finish())
     }
@@ -555,35 +580,35 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         XCTAssertTrue(childChannel.isActive)
 
         // Now we close it several times. This triggers one RST_STREAM frame. The channel will not be closed at this time.
-        var firstClosed = false
-        var secondClosed = false
-        var thirdClosed = false
+        let firstClosed = ManagedAtomic<Bool>(false)
+        let secondClosed = ManagedAtomic<Bool>(false)
+        let thirdClosed = ManagedAtomic<Bool>(false)
         childChannel.close().whenComplete { _ in
-            XCTAssertFalse(firstClosed)
-            XCTAssertFalse(secondClosed)
-            XCTAssertFalse(thirdClosed)
-            firstClosed = true
+            XCTAssertFalse(firstClosed.load(ordering: .sequentiallyConsistent))
+            XCTAssertFalse(secondClosed.load(ordering: .sequentiallyConsistent))
+            XCTAssertFalse(thirdClosed.load(ordering: .sequentiallyConsistent))
+            firstClosed.store(true, ordering: .sequentiallyConsistent)
         }
         childChannel.close().whenComplete { _ in
-            XCTAssertTrue(firstClosed)
-            XCTAssertFalse(secondClosed)
-            XCTAssertFalse(thirdClosed)
-            secondClosed = true
+            XCTAssertTrue(firstClosed.load(ordering: .sequentiallyConsistent))
+            XCTAssertFalse(secondClosed.load(ordering: .sequentiallyConsistent))
+            XCTAssertFalse(thirdClosed.load(ordering: .sequentiallyConsistent))
+            secondClosed.store(true, ordering: .sequentiallyConsistent)
         }
         childChannel.close().whenComplete { _ in
-            XCTAssertTrue(firstClosed)
-            XCTAssertTrue(secondClosed)
-            XCTAssertFalse(thirdClosed)
-            thirdClosed = true
+            XCTAssertTrue(firstClosed.load(ordering: .sequentiallyConsistent))
+            XCTAssertTrue(secondClosed.load(ordering: .sequentiallyConsistent))
+            XCTAssertFalse(thirdClosed.load(ordering: .sequentiallyConsistent))
+            thirdClosed.store(true, ordering: .sequentiallyConsistent)
         }
         XCTAssertEqual(frameReceiver.flushedWrites.count, 1)
         frameReceiver.flushedWrites[0].assertRstStreamFrame(streamID: streamID, errorCode: .cancel)
-        XCTAssertFalse(thirdClosed)
+        XCTAssertFalse(thirdClosed.load(ordering: .sequentiallyConsistent))
 
         // Now send the stream closed event. This will satisfy the close promise.
         let userEvent = StreamClosedEvent(streamID: streamID, reason: .cancel)
         self.channel.pipeline.fireUserInboundEventTriggered(userEvent)
-        XCTAssertTrue(thirdClosed)
+        XCTAssertTrue(thirdClosed.load(ordering: .sequentiallyConsistent))
 
         XCTAssertNoThrow(try self.channel.finish())
     }
@@ -612,16 +637,24 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         XCTAssertTrue(childChannel.isActive)
 
         // Now we close it. This triggers a RST_STREAM frame. The channel will not be closed at this time.
-        var closeError: Error? = nil
-        childChannel.close().whenFailure { closeError = $0 }
+        let closeError = NIOLockedValueBox<Error?>(nil)
+        childChannel.close().whenFailure { error in
+            closeError.withLockedValue { closeError in
+                closeError = error
+            }
+        }
         XCTAssertEqual(frameReceiver.flushedWrites.count, 1)
         frameReceiver.flushedWrites[0].assertRstStreamFrame(streamID: streamID, errorCode: .cancel)
-        XCTAssertNil(closeError)
+        closeError.withLockedValue { closeError in
+            XCTAssertNil(closeError)
+        }
 
         // Now send the stream closed event. This will fail the close promise.
         let userEvent = StreamClosedEvent(streamID: streamID, reason: .cancel)
         self.channel.pipeline.fireUserInboundEventTriggered(userEvent)
-        XCTAssertEqual(closeError as? NIOHTTP2Errors.StreamClosed, NIOHTTP2Errors.StreamClosed(streamID: streamID, errorCode: .cancel))
+        closeError.withLockedValue { closeError in
+            XCTAssertEqual(closeError as? NIOHTTP2Errors.StreamClosed, NIOHTTP2Errors.StreamClosed(streamID: streamID, errorCode: .cancel))
+        }
 
         XCTAssertNoThrow(try self.channel.finish())
     }
@@ -717,10 +750,10 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
 
         // Ok, fail the setup promise. This should deliver a RST_STREAM frame, but not yet close the channel.
         // The channel should, however, be inactive.
-        var channelClosed = false
-        childChannel.closeFuture.whenComplete { _ in channelClosed = true }
+        let channelClosed = ManagedAtomic<Bool>(false)
+        childChannel.closeFuture.whenComplete { _ in channelClosed.store(true, ordering: .sequentiallyConsistent) }
         XCTAssertEqual(writeRecorder.flushedWrites.count, 0)
-        XCTAssertFalse(channelClosed)
+        XCTAssertFalse(channelClosed.load(ordering: .sequentiallyConsistent))
 
         setupCompletePromise.fail(MyError())
         self.channel.assertNoFramesReceived()
@@ -740,20 +773,22 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         XCTAssertEqual(frameRecorder.receivedFrames.count, 0)
         XCTAssertFalse(childChannel.isActive)
         XCTAssertEqual(writeRecorder.flushedWrites.count, 1)
-        XCTAssertFalse(channelClosed)
+        XCTAssertFalse(channelClosed.load(ordering: .sequentiallyConsistent))
 
         (childChannel.eventLoop as! EmbeddedEventLoop).run()
-        XCTAssertTrue(channelClosed)
+        XCTAssertTrue(channelClosed.load(ordering: .sequentiallyConsistent))
 
         XCTAssertNoThrow(try self.channel.finish())
     }
 
     @available(*, deprecated, message: "Deprecated so deprecated functionality can be tested without warnings")
-    func testFlushingOneChannelDoesntFlushThemAll() throws {
+    @available(macOS 10.15, iOS 13.0, watchOS 6.0, tvOS 13.0, *)
+    func testFlushingOneChannelDoesntFlushThemAll() async throws {
         let writeTracker = FrameWriteRecorder()
-        var channels: [Channel] = []
+
+        let (channelsStream, channelsContinuation) = AsyncStream.makeStream(of: Channel.self)
         let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { (channel, _) in
-            channels.append(channel)
+            channelsContinuation.yield(channel)
             return channel.eventLoop.makeSucceededFuture(())
         }
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(writeTracker).wait())
@@ -769,23 +804,25 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
             XCTAssertNoThrow(try self.channel.writeInbound(frame))
             self.activateStream(streamID)
         }
-        XCTAssertEqual(channels.count, 2)
+
+        var streamChannelIterator = channelsStream.makeAsyncIterator()
+        let firstStreamChannel = await streamChannelIterator.next()!
+        let secondStreamChannel = await streamChannelIterator.next()!
 
         // We will now write a headers frame to each channel. Neither frame should be written to the connection. To verify this
         // we will flush the parent channel.
-        for (idx, streamID) in [firstStreamID, secondStreamID].enumerated() {
-            let frame = HTTP2Frame(streamID: streamID, payload: .headers(.init(headers: HPACKHeaders())))
-            channels[idx].write(frame, promise: nil)
-        }
+
+        firstStreamChannel.write(HTTP2Frame(streamID: firstStreamID, payload: .headers(.init(headers: HPACKHeaders()))), promise: nil)
+        secondStreamChannel.write(HTTP2Frame(streamID: secondStreamID, payload: .headers(.init(headers: HPACKHeaders()))), promise: nil)
         self.channel.flush()
         XCTAssertEqual(writeTracker.flushedWrites.count, 0)
 
         // Now we're going to flush only the first child channel. This should cause one flushed write.
-        channels[0].flush()
+        firstStreamChannel.flush()
         XCTAssertEqual(writeTracker.flushedWrites.count, 1)
 
         // Now the other.
-        channels[1].flush()
+        secondStreamChannel.flush()
         XCTAssertEqual(writeTracker.flushedWrites.count, 2)
 
         XCTAssertNoThrow(try self.channel.finish())
@@ -793,9 +830,9 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
 
     @available(*, deprecated, message: "Deprecated so deprecated functionality can be tested without warnings")
     func testUnflushedWritesFailOnClose() throws {
-        var childChannel: Channel? = nil
+        let childChannelPromise = self.channel.eventLoop.makePromise(of: Channel.self)
         let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { (channel, _) in
-            childChannel = channel
+            childChannelPromise.succeed(channel)
             return channel.eventLoop.makeSucceededFuture(())
         }
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(multiplexer).wait())
@@ -805,28 +842,35 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         let frame = HTTP2Frame(streamID: streamID, payload: .headers(.init(headers: HPACKHeaders(), endStream: true)))
         XCTAssertNoThrow(try self.channel.writeInbound(frame))
         XCTAssertNotNil(channel)
+        let childChannel = try childChannelPromise.futureResult.wait()
 
         // We will now write a headers frame to the channel, but don't flush it.
-        var writeError: Error? = nil
+        let writeError = NIOLockedValueBox<Error?>(nil)
         let responseFrame = HTTP2Frame(streamID: streamID, payload: .headers(.init(headers: HPACKHeaders())))
-        childChannel!.write(responseFrame).whenFailure {
-            writeError = $0
+        childChannel.write(responseFrame).whenFailure { error in
+            writeError.withLockedValue { writeError in
+                writeError = error
+            }
         }
-        XCTAssertNil(writeError)
+        writeError.withLockedValue { writeError in
+            XCTAssertNil(writeError)
+        }
 
         // Now we're going to deliver a normal close to the stream.
         let userEvent = StreamClosedEvent(streamID: streamID, reason: nil)
         self.channel.pipeline.fireUserInboundEventTriggered(userEvent)
-        XCTAssertEqual(writeError as? ChannelError, ChannelError.eof)
+        writeError.withLockedValue { writeError in
+            XCTAssertEqual(writeError as? ChannelError, ChannelError.eof)
+        }
 
         XCTAssertNoThrow(try self.channel.finish())
     }
 
     @available(*, deprecated, message: "Deprecated so deprecated functionality can be tested without warnings")
     func testUnflushedWritesFailOnError() throws {
-        var childChannel: Channel? = nil
+        let childChannelPromise = self.channel.eventLoop.makePromise(of: Channel.self)
         let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { (channel, _) in
-            childChannel = channel
+            childChannelPromise.succeed(channel)
             return channel.eventLoop.makeSucceededFuture(())
         }
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(multiplexer).wait())
@@ -836,28 +880,36 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         let frame = HTTP2Frame(streamID: streamID, payload: .headers(.init(headers: HPACKHeaders(), endStream: true)))
         XCTAssertNoThrow(try self.channel.writeInbound(frame))
         XCTAssertNotNil(channel)
+        let childChannel = try childChannelPromise.futureResult.wait()
 
         // We will now write a headers frame to the channel, but don't flush it.
-        var writeError: Error? = nil
+        let writeError = NIOLockedValueBox<Error?>(nil)
         let responseFrame = HTTP2Frame(streamID: streamID, payload: .headers(.init(headers: HPACKHeaders())))
-        childChannel!.write(responseFrame).whenFailure {
-            writeError = $0
+        childChannel.write(responseFrame).whenFailure { error in
+            writeError.withLockedValue { writeError in
+                writeError = error
+            }
         }
-        XCTAssertNil(writeError)
+        writeError.withLockedValue { writeError in
+            XCTAssertNil(writeError)
+        }
 
         // Now we're going to deliver a normal close to the stream.
         let userEvent = StreamClosedEvent(streamID: streamID, reason: .cancel)
         self.channel.pipeline.fireUserInboundEventTriggered(userEvent)
-        XCTAssertEqual(writeError as? NIOHTTP2Errors.StreamClosed, NIOHTTP2Errors.StreamClosed(streamID: streamID, errorCode: .cancel))
+
+        writeError.withLockedValue { writeError in
+            XCTAssertEqual(writeError as? NIOHTTP2Errors.StreamClosed, NIOHTTP2Errors.StreamClosed(streamID: streamID, errorCode: .cancel))
+        }
 
         XCTAssertNoThrow(try self.channel.finish())
     }
 
     @available(*, deprecated, message: "Deprecated so deprecated functionality can be tested without warnings")
     func testWritesFailOnClosedStreamChannels() throws {
-        var childChannel: Channel? = nil
+        let childChannelPromise = self.channel.eventLoop.makePromise(of: Channel.self)
         let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { (channel, _) in
-            childChannel = channel
+            childChannelPromise.succeed(channel)
             return channel.eventLoop.makeSucceededFuture(())
         }
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(multiplexer).wait())
@@ -872,23 +924,29 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         let userEvent = StreamClosedEvent(streamID: streamID, reason: nil)
         self.channel.pipeline.fireUserInboundEventTriggered(userEvent)
 
+        let childChannel = try childChannelPromise.futureResult.wait()
+
         // We will now write a headers frame to the channel. This should fail immediately.
-        var writeError: Error? = nil
+        let writeError = NIOLockedValueBox<Error?>(nil)
         let responseFrame = HTTP2Frame(streamID: streamID, payload: .headers(.init(headers: HPACKHeaders())))
-        childChannel!.write(responseFrame).whenFailure {
-            writeError = $0
+        childChannel.write(responseFrame).whenFailure { error in
+            writeError.withLockedValue { writeError in
+                writeError = error
+            }
         }
-        XCTAssertEqual(writeError as? ChannelError, ChannelError.ioOnClosedChannel)
+        writeError.withLockedValue { writeError in
+            XCTAssertEqual(writeError as? ChannelError, ChannelError.ioOnClosedChannel)
+        }
 
         XCTAssertNoThrow(try self.channel.finish())
     }
 
     @available(*, deprecated, message: "Deprecated so deprecated functionality can be tested without warnings")
     func testReadPullsInAllFrames() throws {
-        var childChannel: Channel? = nil
+        let childChannelPromise = self.channel.eventLoop.makePromise(of: Channel.self)
         let frameRecorder = InboundFrameRecorder()
         let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { (channel, _) -> EventLoopFuture<Void> in
-            childChannel = channel
+            childChannelPromise.succeed(channel)
 
             // We're going to disable autoRead on this channel.
             return channel.getOption(ChannelOptions.autoRead).map {
@@ -908,12 +966,13 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
 
         XCTAssertNoThrow(try self.channel.connect(to: SocketAddress(unixDomainSocketPath: "/whatever"), promise: nil))
 
+
         // Let's open a stream.
         let streamID = HTTP2StreamID(1)
         let frame = HTTP2Frame(streamID: streamID, payload: .headers(.init(headers: HPACKHeaders())))
         XCTAssertNoThrow(try self.channel.writeInbound(frame))
         self.activateStream(1)
-        XCTAssertNotNil(childChannel)
+        let childChannel = try childChannelPromise.futureResult.wait()
 
         // Now we're going to deliver 5 data frames for this stream.
         var buffer = self.channel.allocator.buffer(capacity: 12)
@@ -927,7 +986,7 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         XCTAssertEqual(frameRecorder.receivedFrames.count, 0)
 
         // We'll call read() on the child channel.
-        childChannel!.read()
+        childChannel.read()
 
         // All frames should now have been delivered.
         XCTAssertEqual(frameRecorder.receivedFrames.count, 6)
@@ -940,14 +999,14 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
     }
 
     @available(*, deprecated, message: "Deprecated so deprecated functionality can be tested without warnings")
-    func testReadIsPerChannel() throws {
+    @available(macOS 10.15, iOS 13.0, watchOS 6.0, tvOS 13.0, *)
+    func testReadIsPerChannel() async throws {
         let firstStreamID = HTTP2StreamID(1)
         let secondStreamID = HTTP2StreamID(3)
-        var frameRecorders: [HTTP2StreamID: InboundFrameRecorder] = [:]
-
+        let (channelsStream, channelsContinuation) = AsyncStream.makeStream(of: Channel.self)
         let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { (channel, streamID) -> EventLoopFuture<Void> in
             let recorder = InboundFrameRecorder()
-            frameRecorders[streamID] = recorder
+            channelsContinuation.yield(channel)
 
             // Disable autoRead on the first channel.
             return channel.setOption(ChannelOptions.autoRead, value: streamID != firstStreamID).flatMap {
@@ -964,11 +1023,18 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
             XCTAssertNoThrow(try self.channel.writeInbound(frame))
             self.activateStream(streamID)
         }
-        XCTAssertEqual(frameRecorders.count, 2)
+
+        var streamChannelIterator = channelsStream.makeAsyncIterator()
+        let firstStreamChannel = await streamChannelIterator.next()!
+        let secondStreamChannel = await streamChannelIterator.next()!
 
         // Stream 1 should not have received a frame, stream 3 should.
-        XCTAssertEqual(frameRecorders[firstStreamID]!.receivedFrames.count, 0)
-        XCTAssertEqual(frameRecorders[secondStreamID]!.receivedFrames.count, 1)
+        try await firstStreamChannel.pipeline.handler(type: InboundFrameRecorder.self).map { recorder in
+            XCTAssertEqual(recorder.receivedFrames.count, 0)
+        }.get()
+        try await secondStreamChannel.pipeline.handler(type: InboundFrameRecorder.self).map { recorder in
+            XCTAssertEqual(recorder.receivedFrames.count, 1)
+        }.get()
 
         // Deliver a DATA frame to each stream, which should also have gone into stream 3 but not stream 1.
         var buffer = self.channel.allocator.buffer(capacity: 12)
@@ -978,19 +1044,24 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
             XCTAssertNoThrow(try self.channel.writeInbound(frame))
         }
 
+        try await firstStreamChannel.pipeline.handler(type: InboundFrameRecorder.self).map { recorder in
+            XCTAssertEqual(recorder.receivedFrames.count, 0)
+        }.get()
+
         // Stream 1 should not have received a frame, stream 3 should.
-        XCTAssertEqual(frameRecorders[firstStreamID]!.receivedFrames.count, 0)
-        XCTAssertEqual(frameRecorders[secondStreamID]!.receivedFrames.count, 2)
+        try await secondStreamChannel.pipeline.handler(type: InboundFrameRecorder.self).map { recorder in
+            XCTAssertEqual(recorder.receivedFrames.count, 2)
+        }.get()
 
         XCTAssertNoThrow(try self.channel.finish())
     }
 
     @available(*, deprecated, message: "Deprecated so deprecated functionality can be tested without warnings")
     func testReadWillCauseAutomaticFrameDelivery() throws {
-        var childChannel: Channel? = nil
+        let childChannelPromise = self.channel.eventLoop.makePromise(of: Channel.self)
         let frameRecorder = InboundFrameRecorder()
         let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { (channel, _) -> EventLoopFuture<Void> in
-            childChannel = channel
+            childChannelPromise.succeed(channel)
 
             // We're going to disable autoRead on this channel.
             return channel.setOption(ChannelOptions.autoRead, value: false).flatMap {
@@ -1006,17 +1077,19 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         let frame = HTTP2Frame(streamID: streamID, payload: .headers(.init(headers: HPACKHeaders())))
         XCTAssertNoThrow(try self.channel.writeInbound(frame))
         self.activateStream(streamID)
-        XCTAssertNotNil(childChannel)
+
+
+        let childChannel = try childChannelPromise.futureResult.wait()
 
         // This stream should have seen no frames.
         XCTAssertEqual(frameRecorder.receivedFrames.count, 0)
 
         // Call read, the header frame will come through.
-        childChannel!.read()
+        childChannel.read()
         XCTAssertEqual(frameRecorder.receivedFrames.count, 1)
 
         // Call read again, nothing happens.
-        childChannel!.read()
+        childChannel.read()
         XCTAssertEqual(frameRecorder.receivedFrames.count, 1)
 
         // Now deliver a data frame.
@@ -1037,11 +1110,11 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
 
     @available(*, deprecated, message: "Deprecated so deprecated functionality can be tested without warnings")
     func testReadWithNoPendingDataCausesReadOnParentChannel() throws {
-        var childChannel: Channel? = nil
+        let childChannelPromise = self.channel.eventLoop.makePromise(of: Channel.self)
         let readCounter = ReadCounter()
         let frameRecorder = InboundFrameRecorder()
         let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { (channel, _) -> EventLoopFuture<Void> in
-            childChannel = channel
+            childChannelPromise.succeed(channel)
 
             // We're going to disable autoRead on this channel.
             return channel.setOption(ChannelOptions.autoRead, value: false).flatMap {
@@ -1058,23 +1131,31 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         let frame = HTTP2Frame(streamID: streamID, payload: .headers(.init(headers: HPACKHeaders())))
         XCTAssertNoThrow(try self.channel.writeInbound(frame))
         self.activateStream(streamID)
-        XCTAssertNotNil(childChannel)
+
+
+        let childChannel = try childChannelPromise.futureResult.wait()
 
         // This stream should have seen no frames.
         XCTAssertEqual(frameRecorder.receivedFrames.count, 0)
 
         // There should be no calls to read.
-        XCTAssertEqual(readCounter.readCount, 0)
+        readCounter.readCount.withLockedValue { readCount in
+            XCTAssertEqual(readCount, 0)
+        }
 
         // Call read, the header frame will come through. No calls to read on the parent stream.
-        childChannel!.read()
+        childChannel.read()
         XCTAssertEqual(frameRecorder.receivedFrames.count, 1)
-        XCTAssertEqual(readCounter.readCount, 0)
+        readCounter.readCount.withLockedValue { readCount in
+            XCTAssertEqual(readCount, 0)
+        }
 
         // Call read again, read is called on the parent stream. No frames delivered.
-        childChannel!.read()
+        childChannel.read()
         XCTAssertEqual(frameRecorder.receivedFrames.count, 1)
-        XCTAssertEqual(readCounter.readCount, 1)
+        readCounter.readCount.withLockedValue { readCount in
+            XCTAssertEqual(readCount, 1)
+        }
 
         // Now deliver a data frame.
         var buffer = self.channel.allocator.buffer(capacity: 12)
@@ -1084,33 +1165,41 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
 
         // This frame should have been immediately delivered. No extra call to read.
         XCTAssertEqual(frameRecorder.receivedFrames.count, 2)
-        XCTAssertEqual(readCounter.readCount, 1)
+        readCounter.readCount.withLockedValue { readCount in
+            XCTAssertEqual(readCount, 1)
+        }
 
         // Another call to read issues a read to the parent stream.
-        childChannel!.read()
+        childChannel.read()
         XCTAssertEqual(frameRecorder.receivedFrames.count, 2)
-        XCTAssertEqual(readCounter.readCount, 2)
+        readCounter.readCount.withLockedValue { readCount in
+            XCTAssertEqual(readCount, 2)
+        }
 
         // Another call to read, this time does not issue a read to the parent stream.
-        childChannel!.read()
+        childChannel.read()
         XCTAssertEqual(frameRecorder.receivedFrames.count, 2)
-        XCTAssertEqual(readCounter.readCount, 2)
+        readCounter.readCount.withLockedValue { readCount in
+            XCTAssertEqual(readCount, 2)
+        }
 
         // Delivering two more frames does not cause another call to read, and only one frame
         // is delivered.
         XCTAssertNoThrow(try self.channel.writeInbound(dataFrame))
         XCTAssertNoThrow(try self.channel.writeInbound(dataFrame))
         XCTAssertEqual(frameRecorder.receivedFrames.count, 3)
-        XCTAssertEqual(readCounter.readCount, 2)
+        readCounter.readCount.withLockedValue { readCount in
+            XCTAssertEqual(readCount, 2)
+        }
 
         XCTAssertNoThrow(try self.channel.finish())
     }
 
     @available(*, deprecated, message: "Deprecated so deprecated functionality can be tested without warnings")
     func testHandlersAreRemovedOnClosure() throws {
-        var handlerRemoved = false
+        let handlerRemoved = ManagedAtomic<Bool>(false)
         let handlerRemovedPromise: EventLoopPromise<Void> = self.channel.eventLoop.makePromise()
-        handlerRemovedPromise.futureResult.whenComplete { _ in handlerRemoved = true }
+        handlerRemovedPromise.futureResult.whenComplete { _ in handlerRemoved.store(true, ordering: .sequentiallyConsistent) }
 
         let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { (channel, _) in
             return channel.pipeline.addHandler(HandlerRemovedHandler(removedPromise: handlerRemovedPromise))
@@ -1123,25 +1212,25 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         XCTAssertNoThrow(try self.channel.writeInbound(frame))
 
         // No handlerRemoved so far.
-        XCTAssertFalse(handlerRemoved)
+        XCTAssertFalse(handlerRemoved.load(ordering: .sequentiallyConsistent))
 
         // Now we send the channel a clean exit.
         let event = StreamClosedEvent(streamID: streamID, reason: nil)
         self.channel.pipeline.fireUserInboundEventTriggered(event)
-        XCTAssertFalse(handlerRemoved)
+        XCTAssertFalse(handlerRemoved.load(ordering: .sequentiallyConsistent))
 
         // The handlers will only be removed after we spin the loop.
         (self.channel.eventLoop as! EmbeddedEventLoop).run()
-        XCTAssertTrue(handlerRemoved)
+        XCTAssertTrue(handlerRemoved.load(ordering: .sequentiallyConsistent))
 
         XCTAssertNoThrow(try self.channel.finish())
     }
 
     @available(*, deprecated, message: "Deprecated so deprecated functionality can be tested without warnings")
     func testHandlersAreRemovedOnClosureWithError() throws {
-        var handlerRemoved = false
+        let handlerRemoved = ManagedAtomic<Bool>(false)
         let handlerRemovedPromise: EventLoopPromise<Void> = self.channel.eventLoop.makePromise()
-        handlerRemovedPromise.futureResult.whenComplete { _ in handlerRemoved = true }
+        handlerRemovedPromise.futureResult.whenComplete { _ in handlerRemoved.store(true, ordering: .sequentiallyConsistent) }
 
         let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { (channel, _) in
             return channel.pipeline.addHandler(HandlerRemovedHandler(removedPromise: handlerRemovedPromise))
@@ -1154,16 +1243,16 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         XCTAssertNoThrow(try self.channel.writeInbound(frame))
 
         // No handlerRemoved so far.
-        XCTAssertFalse(handlerRemoved)
+        XCTAssertFalse(handlerRemoved.load(ordering: .sequentiallyConsistent))
 
         // Now we send the channel a clean exit.
         let event = StreamClosedEvent(streamID: streamID, reason: .cancel)
         self.channel.pipeline.fireUserInboundEventTriggered(event)
-        XCTAssertFalse(handlerRemoved)
+        XCTAssertFalse(handlerRemoved.load(ordering: .sequentiallyConsistent))
 
         // The handlers will only be removed after we spin the loop.
         (self.channel.eventLoop as! EmbeddedEventLoop).run()
-        XCTAssertTrue(handlerRemoved)
+        XCTAssertTrue(handlerRemoved.load(ordering: .sequentiallyConsistent))
 
         XCTAssertNoThrow(try self.channel.finish())
     }
@@ -1171,40 +1260,44 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
     @available(*, deprecated, message: "Deprecated so deprecated functionality can be tested without warnings")
     func testCreatingOutboundChannel() throws {
         let configurePromise: EventLoopPromise<Void> = self.channel.eventLoop.makePromise()
-        var createdChannelCount = 0
-        var configuredChannelCount = 0
-        var streamIDs = Array<HTTP2StreamID>()
-        let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { (_, _) in
+        let createdChannelCount = ManagedAtomic<Int>(0)
+        let configuredChannelCount = ManagedAtomic<Int>(0)
+        let streamIDs = NIOLockedValueBox(Array<HTTP2StreamID>())
+        let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { (channel, _) in
             XCTFail("Must not be called")
-            return self.channel.eventLoop.makeFailedFuture(MyError())
+            return channel.eventLoop.makeFailedFuture(MyError())
         }
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(multiplexer).wait())
 
         for _ in 0..<3 {
             let channelPromise: EventLoopPromise<Channel> = self.channel.eventLoop.makePromise()
+
             multiplexer.createStreamChannel(promise: channelPromise) { (channel, streamID) in
-                createdChannelCount += 1
-                streamIDs.append(streamID)
+                createdChannelCount.wrappingIncrement(ordering: .sequentiallyConsistent)
+                streamIDs.withLockedValue { streamIDs in
+                    streamIDs.append(streamID)
+                }
                 return configurePromise.futureResult
             }
+
             channelPromise.futureResult.whenSuccess { _ in
-                configuredChannelCount += 1
+                configuredChannelCount.wrappingIncrement(ordering: .sequentiallyConsistent)
             }
         }
 
-        XCTAssertEqual(createdChannelCount, 0)
-        XCTAssertEqual(configuredChannelCount, 0)
-        XCTAssertEqual(streamIDs.count, 0)
-
         (self.channel.eventLoop as! EmbeddedEventLoop).run()
-        XCTAssertEqual(createdChannelCount, 3)
-        XCTAssertEqual(configuredChannelCount, 0)
-        XCTAssertEqual(streamIDs, [2, 4, 6].map { HTTP2StreamID($0) })
+        XCTAssertEqual(createdChannelCount.load(ordering: .sequentiallyConsistent), 3)
+        XCTAssertEqual(configuredChannelCount.load(ordering: .sequentiallyConsistent), 0)
+        streamIDs.withLockedValue { streamIDs in
+            XCTAssertEqual(streamIDs, [2, 4, 6].map { HTTP2StreamID($0) })
+        }
 
         configurePromise.succeed(())
-        XCTAssertEqual(createdChannelCount, 3)
-        XCTAssertEqual(configuredChannelCount, 3)
-        XCTAssertEqual(streamIDs, [2, 4, 6].map { HTTP2StreamID($0) })
+        XCTAssertEqual(createdChannelCount.load(ordering: .sequentiallyConsistent), 3)
+        XCTAssertEqual(configuredChannelCount.load(ordering: .sequentiallyConsistent), 3)
+        streamIDs.withLockedValue { streamIDs in
+            XCTAssertEqual(streamIDs, [2, 4, 6].map { HTTP2StreamID($0) })
+        }
 
         XCTAssertNoThrow(try self.channel.finish())
     }
@@ -1212,40 +1305,42 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
     @available(*, deprecated, message: "Deprecated so deprecated functionality can be tested without warnings")
     func testCreatingOutboundChannelClient() throws {
         let configurePromise: EventLoopPromise<Void> = self.channel.eventLoop.makePromise()
-        var createdChannelCount = 0
-        var configuredChannelCount = 0
-        var streamIDs = Array<HTTP2StreamID>()
-        let multiplexer = HTTP2StreamMultiplexer(mode: .client, channel: self.channel) { (_, _) in
+        let createdChannelCount = ManagedAtomic<Int>(0)
+        let configuredChannelCount = ManagedAtomic<Int>(0)
+        let streamIDs = NIOLockedValueBox(Array<HTTP2StreamID>())
+        let multiplexer = HTTP2StreamMultiplexer(mode: .client, channel: self.channel) { (channel, _) in
             XCTFail("Must not be called")
-            return self.channel.eventLoop.makeFailedFuture(MyError())
+            return channel.eventLoop.makeFailedFuture(MyError())
         }
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(multiplexer).wait())
 
         for _ in 0..<3 {
             let channelPromise: EventLoopPromise<Channel> = self.channel.eventLoop.makePromise()
             multiplexer.createStreamChannel(promise: channelPromise) { (channel, streamID) in
-                createdChannelCount += 1
-                streamIDs.append(streamID)
+                createdChannelCount.wrappingIncrement(ordering: .sequentiallyConsistent)
+                streamIDs.withLockedValue { streamIDs in
+                    streamIDs.append(streamID)
+                }
                 return configurePromise.futureResult
             }
             channelPromise.futureResult.whenSuccess { _ in
-                configuredChannelCount += 1
+                configuredChannelCount.wrappingIncrement(ordering: .sequentiallyConsistent)
             }
         }
 
-        XCTAssertEqual(createdChannelCount, 0)
-        XCTAssertEqual(configuredChannelCount, 0)
-        XCTAssertEqual(streamIDs.count, 0)
-
         (self.channel.eventLoop as! EmbeddedEventLoop).run()
-        XCTAssertEqual(createdChannelCount, 3)
-        XCTAssertEqual(configuredChannelCount, 0)
-        XCTAssertEqual(streamIDs, [1, 3, 5].map { HTTP2StreamID($0) })
+        XCTAssertEqual(createdChannelCount.load(ordering: .sequentiallyConsistent), 3)
+        XCTAssertEqual(configuredChannelCount.load(ordering: .sequentiallyConsistent), 0)
+        streamIDs.withLockedValue { streamIDs in
+            XCTAssertEqual(streamIDs, [1, 3, 5].map { HTTP2StreamID($0) })
+        }
 
         configurePromise.succeed(())
-        XCTAssertEqual(createdChannelCount, 3)
-        XCTAssertEqual(configuredChannelCount, 3)
-        XCTAssertEqual(streamIDs, [1, 3, 5].map { HTTP2StreamID($0) })
+        XCTAssertEqual(createdChannelCount.load(ordering: .sequentiallyConsistent), 3)
+        XCTAssertEqual(configuredChannelCount.load(ordering: .sequentiallyConsistent), 3)
+        streamIDs.withLockedValue { streamIDs in
+            XCTAssertEqual(streamIDs, [1, 3, 5].map { HTTP2StreamID($0) })
+        }
 
         XCTAssertNoThrow(try self.channel.finish())
     }
@@ -1254,27 +1349,30 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
     func testWritesOnCreatedChannelAreDelayed() throws {
         let configurePromise: EventLoopPromise<Void> = self.channel.eventLoop.makePromise()
         let writeRecorder = FrameWriteRecorder()
-        var childChannel: Channel? = nil
-        var childStreamID: HTTP2StreamID? = nil
+        let childChannelPromise = self.channel.eventLoop.makePromise(of: Channel.self)
+        let childStreamIDPromise = self.channel.eventLoop.makePromise(of: HTTP2StreamID.self)
 
         XCTAssertNoThrow(try self.channel.connect(to: SocketAddress(unixDomainSocketPath: "/whatever"), promise: nil))
 
-        let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { (_, _) in
+        let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { (channel, _) in
             XCTFail("Must not be called")
-            return self.channel.eventLoop.makeFailedFuture(MyError())
+            return channel.eventLoop.makeFailedFuture(MyError())
         }
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(writeRecorder).wait())
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(multiplexer).wait())
         multiplexer.createStreamChannel(promise: nil) { (channel, streamID) in
-            childChannel = channel
-            childStreamID = streamID
+            childChannelPromise.succeed(channel)
+            childStreamIDPromise.succeed(streamID)
             return configurePromise.futureResult
         }
         (self.channel.eventLoop as! EmbeddedEventLoop).run()
-        XCTAssertNotNil(childChannel)
-        XCTAssertNotNil(childStreamID)
 
-        childChannel!.writeAndFlush(HTTP2Frame(streamID: childStreamID!, payload: .headers(.init(headers: HPACKHeaders()))), promise: nil)
+
+        let childChannel = try childChannelPromise.futureResult.wait()
+        let childStreamID = try childStreamIDPromise.futureResult.wait()
+
+        childChannel.writeAndFlush(HTTP2Frame(streamID: childStreamID, payload: .headers(.init(headers: HPACKHeaders()))), promise: nil)
+
         XCTAssertEqual(writeRecorder.flushedWrites.count, 0)
 
         configurePromise.succeed(())
@@ -1287,28 +1385,41 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
     @available(*, deprecated, message: "Deprecated so deprecated functionality can be tested without warnings")
     func testWritesAreCancelledOnFailingInitializer() throws {
         let configurePromise: EventLoopPromise<Void> = self.channel.eventLoop.makePromise()
-        var childChannel: Channel? = nil
-        var childStreamID: HTTP2StreamID? = nil
+        let childChannelPromise = self.channel.eventLoop.makePromise(of: Channel.self)
+        let childStreamIDPromise = self.channel.eventLoop.makePromise(of: HTTP2StreamID.self)
 
-        let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { (_, _) in
+        let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { (channel, _) in
             XCTFail("Must not be called")
-            return self.channel.eventLoop.makeFailedFuture(MyError())
+            return channel.eventLoop.makeFailedFuture(MyError())
         }
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(multiplexer).wait())
         multiplexer.createStreamChannel(promise: nil) { (channel, streamID) in
-            childChannel = channel
-            childStreamID = streamID
+            childChannelPromise.succeed(channel)
+            childStreamIDPromise.succeed(streamID)
             return configurePromise.futureResult
         }
         (self.channel.eventLoop as! EmbeddedEventLoop).run()
 
-        var writeError: Error? = nil
-        childChannel!.writeAndFlush(HTTP2Frame(streamID: childStreamID!, payload: .headers(.init(headers: HPACKHeaders())))).whenFailure { writeError = $0 }
-        XCTAssertNil(writeError)
+        let childChannel = try childChannelPromise.futureResult.wait()
+        let childStreamID = try childStreamIDPromise.futureResult.wait()
+
+        let writeError = NIOLockedValueBox<Error?>(nil)
+
+        childChannel.writeAndFlush(HTTP2Frame(streamID: childStreamID, payload: .headers(.init(headers: HPACKHeaders())))).whenFailure { error in
+            writeError.withLockedValue { writeError in
+                writeError = error
+            }
+        }
+
+        writeError.withLockedValue { writeError in
+            XCTAssertNil(writeError)
+        }
 
         configurePromise.fail(MyError())
-        XCTAssertNotNil(writeError)
-        XCTAssertTrue(writeError is MyError)
+        writeError.withLockedValue { writeError in
+            XCTAssertNotNil(writeError)
+            XCTAssertTrue(writeError is MyError)
+        }
 
         XCTAssertNoThrow(try self.channel.finish())
     }
@@ -1318,9 +1429,9 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         let configurePromise: EventLoopPromise<Void> = self.channel.eventLoop.makePromise()
         let writeRecorder = FrameWriteRecorder()
 
-        let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { (_, _) in
+        let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { (channel, _) in
             XCTFail("Must not be called")
-            return self.channel.eventLoop.makeFailedFuture(MyError())
+            return channel.eventLoop.makeFailedFuture(MyError())
         }
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(writeRecorder).wait())
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(multiplexer).wait())
@@ -1337,60 +1448,60 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
 
     @available(*, deprecated, message: "Deprecated so deprecated functionality can be tested without warnings")
     func testCreatedChildChannelDoesNotActivateEarly() throws {
-        var activated = false
+        let activated = ManagedAtomic<Bool>(false)
 
         let activePromise: EventLoopPromise<Void> = self.channel.eventLoop.makePromise()
-        let activeRecorder = ActiveHandler(activatedPromise: activePromise)
         activePromise.futureResult.map {
-            activated = true
+            activated.store(true, ordering: .sequentiallyConsistent)
         }.whenFailure { (_: Error) in
             XCTFail("Activation promise must not fail")
         }
 
-        let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { (_, _) in
+        let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { (channel, _) in
             XCTFail("Must not be called")
-            return self.channel.eventLoop.makeFailedFuture(MyError())
+            return channel.eventLoop.makeFailedFuture(MyError())
         }
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(multiplexer).wait())
         multiplexer.createStreamChannel(promise: nil) { (channel, streamID) in
+            let activeRecorder = ActiveHandler(activatedPromise: activePromise)
             return channel.pipeline.addHandler(activeRecorder)
         }
         (self.channel.eventLoop as! EmbeddedEventLoop).run()
-        XCTAssertFalse(activated)
+        XCTAssertFalse(activated.load(ordering: .sequentiallyConsistent))
 
         XCTAssertNoThrow(try self.channel.connect(to: SocketAddress(unixDomainSocketPath: "/whatever"), promise: nil))
 
-        XCTAssertTrue(activated)
+        XCTAssertTrue(activated.load(ordering: .sequentiallyConsistent))
 
         XCTAssertNoThrow(try self.channel.finish())
     }
 
     @available(*, deprecated, message: "Deprecated so deprecated functionality can be tested without warnings")
     func testCreatedChildChannelActivatesIfParentIsActive() throws {
-        var activated = false
+        let activated = ManagedAtomic<Bool>(false)
 
         let activePromise: EventLoopPromise<Void> = self.channel.eventLoop.makePromise()
-        let activeRecorder = ActiveHandler(activatedPromise: activePromise)
         activePromise.futureResult.map {
-            activated = true
+            activated.store(true, ordering: .sequentiallyConsistent)
         }.whenFailure { (_: Error) in
             XCTFail("Activation promise must not fail")
         }
 
-        let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { (_, _) in
+        let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { (channel, _) in
             XCTFail("Must not be called")
-            return self.channel.eventLoop.makeFailedFuture(MyError())
+            return channel.eventLoop.makeFailedFuture(MyError())
         }
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(multiplexer).wait())
 
         XCTAssertNoThrow(try self.channel.connect(to: SocketAddress(ipAddress: "127.0.0.1", port: 8765)).wait())
-        XCTAssertFalse(activated)
+        XCTAssertFalse(activated.load(ordering: .sequentiallyConsistent))
 
         multiplexer.createStreamChannel(promise: nil) { (channel, streamID) in
+            let activeRecorder = ActiveHandler(activatedPromise: activePromise)
             return channel.pipeline.addHandler(activeRecorder)
         }
         (self.channel.eventLoop as! EmbeddedEventLoop).run()
-        XCTAssertTrue(activated)
+        XCTAssertTrue(activated.load(ordering: .sequentiallyConsistent))
 
         XCTAssertNoThrow(try self.channel.finish())
     }
@@ -1399,28 +1510,28 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
     func testInitiatedChildChannelActivates() throws {
         XCTAssertNoThrow(try self.channel.connect(to: SocketAddress(unixDomainSocketPath: "/whatever"), promise: nil))
 
-        var activated = false
+        let activated = ManagedAtomic<Bool>(false)
 
         let activePromise: EventLoopPromise<Void> = self.channel.eventLoop.makePromise()
-        let activeRecorder = ActiveHandler(activatedPromise: activePromise)
         activePromise.futureResult.map {
-            activated = true
+            activated.store(true, ordering: .sequentiallyConsistent)
         }.whenFailure { (_: Error) in
             XCTFail("Activation promise must not fail")
         }
 
         let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { (channel, _) in
+            let activeRecorder = ActiveHandler(activatedPromise: activePromise)
             return channel.pipeline.addHandler(activeRecorder)
         }
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(multiplexer).wait())
         self.channel.pipeline.fireChannelActive()
 
         // Open a new stream.
-        XCTAssertFalse(activated)
+        XCTAssertFalse(activated.load(ordering: .sequentiallyConsistent))
         let streamID = HTTP2StreamID(1)
         let frame = HTTP2Frame(streamID: streamID, payload: .headers(.init(headers: HPACKHeaders())))
         XCTAssertNoThrow(try self.channel.writeInbound(frame))
-        XCTAssertTrue(activated)
+        XCTAssertTrue(activated.load(ordering: .sequentiallyConsistent))
 
         XCTAssertNoThrow(try self.channel.finish())
     }
@@ -1440,46 +1551,48 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
     func testMultiplexerForwardsActiveToParent() throws {
         self.channel.addNoOpMultiplexer(mode: .client)
 
-        var didActivate = false
+        let didActivate = ManagedAtomic<Bool>(false)
 
         let activePromise = self.channel.eventLoop.makePromise(of: Void.self)
         activePromise.futureResult.whenSuccess {
-            didActivate = true
+            didActivate.store(true, ordering: .sequentiallyConsistent)
         }
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(ActiveHandler(activatedPromise: activePromise)).wait())
         XCTAssertNoThrow(try self.channel.connect(to: SocketAddress(unixDomainSocketPath: "/nothing")).wait())
-        XCTAssertTrue(didActivate)
-        
+        XCTAssertTrue(didActivate.load(ordering: .sequentiallyConsistent))
+
         XCTAssertNoThrow(try self.channel.finish())
     }
 
     @available(*, deprecated, message: "Deprecated so deprecated functionality can be tested without warnings")
     func testCreatedChildChannelCanBeClosedImmediately() throws {
-        var closed = false
+        let closed = ManagedAtomic<Bool>(false)
 
         let multiplexer = HTTP2StreamMultiplexer(mode: .client, channel: self.channel) { (channel, _) in
             XCTFail("Must not be called")
-            return self.channel.eventLoop.makeFailedFuture(MyError())
+            return channel.eventLoop.makeFailedFuture(MyError())
         }
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(multiplexer).wait())
 
-        XCTAssertFalse(closed)
+        XCTAssertFalse(closed.load(ordering: .sequentiallyConsistent))
         multiplexer.createStreamChannel(promise: nil) { (channel, streamID) in
-            channel.close().whenComplete { _ in closed = true }
+            channel.close().whenComplete { _ in
+                closed.store(true, ordering: .sequentiallyConsistent)
+            }
             return channel.eventLoop.makeSucceededFuture(())
         }
         self.channel.embeddedEventLoop.run()
-        XCTAssertTrue(closed)
+        XCTAssertTrue(closed.load(ordering: .sequentiallyConsistent))
         XCTAssertNoThrow(XCTAssertTrue(try self.channel.finish().isClean))
     }
 
     @available(*, deprecated, message: "Deprecated so deprecated functionality can be tested without warnings")
     func testCreatedChildChannelCanBeClosedBeforeWritingHeaders() throws {
-        var closed = false
+        let closed = ManagedAtomic<Bool>(false)
 
         let multiplexer = HTTP2StreamMultiplexer(mode: .client, channel: self.channel) { (channel, _) in
             XCTFail("Must not be called")
-            return self.channel.eventLoop.makeFailedFuture(MyError())
+            return channel.eventLoop.makeFailedFuture(MyError())
         }
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(multiplexer).wait())
 
@@ -1490,48 +1603,52 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         self.channel.embeddedEventLoop.run()
 
         let child = try assertNoThrowWithValue(channelPromise.futureResult.wait())
-        child.closeFuture.whenComplete { _ in closed = true }
+        child.closeFuture.whenComplete { _ in
+            closed.store(true, ordering: .sequentiallyConsistent)
+        }
 
-        XCTAssertFalse(closed)
+        XCTAssertFalse(closed.load(ordering: .sequentiallyConsistent))
         child.close(promise: nil)
         self.channel.embeddedEventLoop.run()
-        XCTAssertTrue(closed)
+        XCTAssertTrue(closed.load(ordering: .sequentiallyConsistent))
         XCTAssertNoThrow(XCTAssertTrue(try self.channel.finish().isClean))
     }
 
     @available(*, deprecated, message: "Deprecated so deprecated functionality can be tested without warnings")
     func testCreatedChildChannelCanBeClosedImmediatelyWhenBaseIsActive() throws {
-        var closed = false
+        let closed = ManagedAtomic<Bool>(false)
 
         // We need to activate the underlying channel here.
         XCTAssertNoThrow(try self.channel.connect(to: SocketAddress(ipAddress: "127.0.0.1", port: 80)).wait())
 
         let multiplexer = HTTP2StreamMultiplexer(mode: .client, channel: self.channel) { (channel, _) in
             XCTFail("Must not be called")
-            return self.channel.eventLoop.makeFailedFuture(MyError())
+            return channel.eventLoop.makeFailedFuture(MyError())
         }
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(multiplexer).wait())
 
-        XCTAssertFalse(closed)
+        XCTAssertFalse(closed.load(ordering: .sequentiallyConsistent))
         multiplexer.createStreamChannel(promise: nil) { (channel, streamID) in
-            channel.close().whenComplete { _ in closed = true }
+            channel.close().whenComplete { _ in
+                closed.store(true, ordering: .sequentiallyConsistent)
+            }
             return channel.eventLoop.makeSucceededFuture(())
         }
         self.channel.embeddedEventLoop.run()
-        XCTAssertTrue(closed)
+        XCTAssertTrue(closed.load(ordering: .sequentiallyConsistent))
         XCTAssertNoThrow(XCTAssertTrue(try self.channel.finish().isClean))
     }
 
     @available(*, deprecated, message: "Deprecated so deprecated functionality can be tested without warnings")
     func testCreatedChildChannelCanBeClosedBeforeWritingHeadersWhenBaseIsActive() throws {
-        var closed = false
+        let closed = ManagedAtomic<Bool>(false)
 
         // We need to activate the underlying channel here.
         XCTAssertNoThrow(try self.channel.connect(to: SocketAddress(ipAddress: "127.0.0.1", port: 80)).wait())
 
         let multiplexer = HTTP2StreamMultiplexer(mode: .client, channel: self.channel) { (channel, _) in
             XCTFail("Must not be called")
-            return self.channel.eventLoop.makeFailedFuture(MyError())
+            return channel.eventLoop.makeFailedFuture(MyError())
         }
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(multiplexer).wait())
 
@@ -1542,12 +1659,14 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         self.channel.embeddedEventLoop.run()
 
         let child = try assertNoThrowWithValue(channelPromise.futureResult.wait())
-        child.closeFuture.whenComplete { _ in closed = true }
+        child.closeFuture.whenComplete { _ in
+            closed.store(true, ordering: .sequentiallyConsistent)
+        }
 
-        XCTAssertFalse(closed)
+        XCTAssertFalse(closed.load(ordering: .sequentiallyConsistent))
         child.close(promise: nil)
         self.channel.embeddedEventLoop.run()
-        XCTAssertTrue(closed)
+        XCTAssertTrue(closed.load(ordering: .sequentiallyConsistent))
         XCTAssertNoThrow(XCTAssertTrue(try self.channel.finish().isClean))
     }
 
@@ -1602,7 +1721,9 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(multiplexer).wait())
 
         XCTAssertEqual(frameRecorder.receivedFrames.count, 0)
-        XCTAssertEqual(readCompleteCounter.readCompleteCount, 0)
+        readCompleteCounter.readCompleteCount.withLockedValue { readCompleteCount in
+            XCTAssertEqual(readCompleteCount, 0)
+        }
 
         // Wake up and activate the stream.
         let requestHeaders = HPACKHeaders([(":path", "/"), (":method", "GET"), (":authority", "localhost"), (":scheme", "https")])
@@ -1612,7 +1733,9 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         self.channel.embeddedEventLoop.run()
 
         XCTAssertEqual(frameRecorder.receivedFrames.count, 1)
-        XCTAssertEqual(readCompleteCounter.readCompleteCount, 1)
+        readCompleteCounter.readCompleteCount.withLockedValue { readCompleteCount in
+            XCTAssertEqual(readCompleteCount, 1)
+        }
 
         // Now we're going to send 9 data frames.
         var requestData = self.channel.allocator.buffer(capacity: 1024)
@@ -1625,19 +1748,25 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
 
         // We should have 1 read (the HEADERS), and zero read completes.
         XCTAssertEqual(frameRecorder.receivedFrames.count, 1)
-        XCTAssertEqual(readCompleteCounter.readCompleteCount, 1)
+        readCompleteCounter.readCompleteCount.withLockedValue { readCompleteCount in
+            XCTAssertEqual(readCompleteCount, 1)
+        }
 
         // Fire read complete on the parent and it'll propagate to the child, also firing the reads..
         self.channel.pipeline.fireChannelReadComplete()
 
         // We should have 10 reads, and one read complete.
         XCTAssertEqual(frameRecorder.receivedFrames.count, 10)
-        XCTAssertEqual(readCompleteCounter.readCompleteCount, 2)
+        readCompleteCounter.readCompleteCount.withLockedValue { readCompleteCount in
+            XCTAssertEqual(readCompleteCount, 2)
+        }
 
         // If we fire a new read complete on the parent, the child doesn't see it this time, as it received no frames.
         self.channel.pipeline.fireChannelReadComplete()
         XCTAssertEqual(frameRecorder.receivedFrames.count, 10)
-        XCTAssertEqual(readCompleteCounter.readCompleteCount, 2)
+        readCompleteCounter.readCompleteCount.withLockedValue { readCompleteCount in
+            XCTAssertEqual(readCompleteCount, 2)
+        }
     }
 
     @available(*, deprecated, message: "Deprecated so deprecated functionality can be tested without warnings")
@@ -1658,7 +1787,9 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(multiplexer).wait())
 
         XCTAssertEqual(frameRecorder.receivedFrames.count, 0)
-        XCTAssertEqual(readCompleteCounter.readCompleteCount, 0)
+        readCompleteCounter.readCompleteCount.withLockedValue { readCompleteCount in
+            XCTAssertEqual(readCompleteCount, 0)
+        }
 
         // Wake up and activate the streams.
         let requestHeaders = HPACKHeaders([(":path", "/"), (":method", "GET"), (":authority", "localhost"), (":scheme", "https")])
@@ -1671,12 +1802,16 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         self.channel.embeddedEventLoop.run()
 
         XCTAssertEqual(frameRecorder.receivedFrames.count, 3)
-        XCTAssertEqual(readCompleteCounter.readCompleteCount, 3)
+        readCompleteCounter.readCompleteCount.withLockedValue { readCompleteCount in
+            XCTAssertEqual(readCompleteCount, 3)
+        }
 
         // Firing in readComplete does not cause a readComplete, as no new frames were delivered.
         self.channel.pipeline.fireChannelReadComplete()
         XCTAssertEqual(frameRecorder.receivedFrames.count, 3)
-        XCTAssertEqual(readCompleteCounter.readCompleteCount, 3)
+        readCompleteCounter.readCompleteCount.withLockedValue { readCompleteCount in
+            XCTAssertEqual(readCompleteCount, 3)
+        }
 
         // Now we're going to send a data frame on stream 1.
         var requestData = self.channel.allocator.buffer(capacity: 1024)
@@ -1687,7 +1822,9 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         // We should have 3 reads, and 3 read completes. We don't actually get this frame yet, because
         // we do not do a "fast-delivery" path.
         XCTAssertEqual(frameRecorder.receivedFrames.count, 3)
-        XCTAssertEqual(readCompleteCounter.readCompleteCount, 3)
+        readCompleteCounter.readCompleteCount.withLockedValue { readCompleteCount in
+            XCTAssertEqual(readCompleteCount, 3)
+        }
 
         // Fire read complete on the parent and it'll propagate to the child, but only to the one
         // that saw a frame.
@@ -1695,12 +1832,16 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
 
         // We should have 4 reads, and 4 read completes.
         XCTAssertEqual(frameRecorder.receivedFrames.count, 4)
-        XCTAssertEqual(readCompleteCounter.readCompleteCount, 4)
+        readCompleteCounter.readCompleteCount.withLockedValue { readCompleteCount in
+            XCTAssertEqual(readCompleteCount, 4)
+        }
 
         // If we fire a new read complete on the parent, the children don't see it.
         self.channel.pipeline.fireChannelReadComplete()
         XCTAssertEqual(frameRecorder.receivedFrames.count, 4)
-        XCTAssertEqual(readCompleteCounter.readCompleteCount, 4)
+        readCompleteCounter.readCompleteCount.withLockedValue { readCompleteCount in
+            XCTAssertEqual(readCompleteCount, 4)
+        }
     }
 
     @available(*, deprecated, message: "Deprecated so deprecated functionality can be tested without warnings")
@@ -1708,9 +1849,9 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         let multiplexer = HTTP2StreamMultiplexer(mode: .client,
                                                  channel: self.channel,
                                                  outboundBufferSizeHighWatermark: 100,
-                                                 outboundBufferSizeLowWatermark: 50) { (_, _) in
+                                                 outboundBufferSizeLowWatermark: 50) { (channel, _) in
             XCTFail("Must not be called")
-            return self.channel.eventLoop.makeFailedFuture(MyError())
+            return channel.eventLoop.makeFailedFuture(MyError())
         }
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(multiplexer).wait())
 
@@ -1759,9 +1900,9 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
 
     @available(*, deprecated, message: "Deprecated so deprecated functionality can be tested without warnings")
     func testMultiplexerModifiesStreamChannelWritabilityBasedOnParentChannelWritability() throws {
-        let multiplexer = HTTP2StreamMultiplexer(mode: .client, channel: self.channel) { (_, _) in
+        let multiplexer = HTTP2StreamMultiplexer(mode: .client, channel: self.channel) { (channel, _) in
             XCTFail("Must not be called")
-            return self.channel.eventLoop.makeFailedFuture(MyError())
+            return channel.eventLoop.makeFailedFuture(MyError())
         }
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(multiplexer).wait())
 
@@ -1808,9 +1949,9 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         let multiplexer = HTTP2StreamMultiplexer(mode: .client,
                                                  channel: self.channel,
                                                  outboundBufferSizeHighWatermark: 100,
-                                                 outboundBufferSizeLowWatermark: 50) { (_, _) in
+                                                 outboundBufferSizeLowWatermark: 50) { (channel, _) in
             XCTFail("Must not be called")
-            return self.channel.eventLoop.makeFailedFuture(MyError())
+            return channel.eventLoop.makeFailedFuture(MyError())
         }
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(multiplexer).wait())
 
@@ -1873,9 +2014,9 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
         let multiplexer = HTTP2StreamMultiplexer(mode: .client,
                                                  channel: self.channel,
                                                  outboundBufferSizeHighWatermark: 100,
-                                                 outboundBufferSizeLowWatermark: 50) { (_, _) in
+                                                 outboundBufferSizeLowWatermark: 50) { (channel, _) in
             XCTFail("Must not be called")
-            return self.channel.eventLoop.makeFailedFuture(MyError())
+            return channel.eventLoop.makeFailedFuture(MyError())
         }
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(multiplexer).wait())
 
@@ -1933,7 +2074,7 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
             typealias InboundOut = Element
             typealias OutboundIn = Element
             typealias OutboundOut = Element
-            
+
             let elementsToBufferCount: Int = 1
             var bufferedElements: CircularBuffer<Element> = []
 
@@ -1941,29 +2082,29 @@ final class HTTP2StreamMultiplexerTests: XCTestCase {
                 while let bufferedElement = bufferedElements.popFirst() {
                     context.fireChannelRead(wrapInboundOut(bufferedElement))
                 }
-                
+
                 context.read()
             }
-            
+
             func channelRead(context: ChannelHandlerContext, data: NIOAny) {
                 bufferedElements.append(unwrapInboundIn(data))
-                
+
                 if bufferedElements.count > elementsToBufferCount {
                     context.fireChannelRead(wrapInboundOut(bufferedElements.removeFirst()))
                 }
             }
         }
-        
+
         XCTAssertNoThrow(try self.channel.connect(to: SocketAddress(unixDomainSocketPath: "/whatever"), promise: nil))
         let streamID = HTTP2StreamID(1)
         let headerFrame = HTTP2Frame(streamID: streamID, payload: .headers(.init(headers: HPACKHeaders())))
         let payloadBuffer = channel.allocator.buffer(capacity: 0)
         let dataFrame = HTTP2Frame(streamID: streamID, payload: .data(.init(data: .byteBuffer(payloadBuffer))))
-        
+
         let multiplexer = HTTP2StreamMultiplexer(mode: .server, channel: self.channel) { channel in
             channel.pipeline.addHandler(FramePayloadExpecter(expectedPayload: [headerFrame.payload, dataFrame.payload]))
         }
-        
+
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(BufferingChannelHandler<HTTP2Frame>()).wait())
         XCTAssertNoThrow(try self.channel.pipeline.addHandler(multiplexer).wait())
 
