@@ -23,9 +23,11 @@ import NIOCore
 ///
 /// Note that while this is a `struct`, this `struct` has _reference semantics_.
 /// The implementation of `Equatable` & `Hashable` on this type reinforces that requirement.
+@usableFromInline
 struct MultiplexerAbstractChannel {
-    private(set) var baseChannel: HTTP2StreamChannel
+    @usableFromInline private(set) var baseChannel: HTTP2StreamChannel
 
+    @usableFromInline
     init(allocator: ByteBufferAllocator,
          parent: Channel,
          multiplexer: HTTP2StreamChannel.OutboundStreamMultiplexer,
@@ -60,10 +62,11 @@ struct MultiplexerAbstractChannel {
 }
 
 extension MultiplexerAbstractChannel {
+    @usableFromInline
     enum InboundStreamStateInitializer {
-        case includesStreamID(((Channel, HTTP2StreamID) -> EventLoopFuture<Void>)?)
-        case excludesStreamID(((Channel) -> EventLoopFuture<Void>)?)
-        case returnsAny(((Channel) -> EventLoopFuture<any Sendable>))
+        case includesStreamID(NIOChannelInitializerWithStreamID?)
+        case excludesStreamID(NIOChannelInitializer?)
+        case returnsAny(NIOChannelInitializerWithOutput<any Sendable>)
     }
 }
 
@@ -73,6 +76,7 @@ extension MultiplexerAbstractChannel {
         return self.baseChannel.streamID
     }
 
+    @usableFromInline
     var channelID: ObjectIdentifier {
         return ObjectIdentifier(self.baseChannel)
     }
@@ -101,7 +105,7 @@ extension MultiplexerAbstractChannel {
         }
     }
 
-    func configureInboundStream(initializer: InboundStreamStateInitializer, promise: EventLoopPromise<Any>?) {
+    func configureInboundStream(initializer: InboundStreamStateInitializer, promise: EventLoopPromise<any Sendable>?) {
         switch initializer {
         case .includesStreamID, .excludesStreamID:
             preconditionFailure("Configuration with a supplied `Any` promise is not supported with this initializer type.")
@@ -111,7 +115,7 @@ extension MultiplexerAbstractChannel {
     }
 
     // legacy `initializer` function signature
-    func configure(initializer: ((Channel, HTTP2StreamID) -> EventLoopFuture<Void>)?, userPromise promise: EventLoopPromise<Channel>?) {
+    func configure(initializer: NIOChannelInitializerWithStreamID?, userPromise promise: EventLoopPromise<Channel>?) {
         self.baseChannel.configure(initializer: initializer, userPromise: promise)
     }
 
@@ -121,7 +125,8 @@ extension MultiplexerAbstractChannel {
     }
 
     // used for async multiplexer
-    func configure(initializer: @escaping NIOChannelInitializerWithOutput<any Sendable>, userPromise promise: EventLoopPromise<Any>?) {
+    @usableFromInline
+    func configure(initializer: @escaping NIOChannelInitializerWithOutput<any Sendable>, userPromise promise: EventLoopPromise<any Sendable>?) {
         self.baseChannel.configure(initializer: initializer, userPromise: promise)
     }
 
@@ -167,12 +172,14 @@ extension MultiplexerAbstractChannel {
 }
 
 extension MultiplexerAbstractChannel: Equatable {
+    @inlinable
     static func ==(lhs: MultiplexerAbstractChannel, rhs: MultiplexerAbstractChannel) -> Bool {
         return lhs.baseChannel === rhs.baseChannel
     }
 }
 
 extension MultiplexerAbstractChannel: Hashable {
+    @inlinable
     func hash(into hasher: inout Hasher) {
         hasher.combine(ObjectIdentifier(self.baseChannel))
     }
