@@ -86,9 +86,8 @@ func sendOneRequest(channel: Channel, multiplexer: HTTP2StreamMultiplexer) throw
                                              ErrorHandler()],
                                             position: .last)
     }
-
-    channel.eventLoop.execute {
-        multiplexer.createStreamChannel(promise: nil, requestStreamInitializer)
+    channel.pipeline.handler(type: HTTP2StreamMultiplexer.self).whenSuccess { multiplexer in
+      multiplexer.createStreamChannel(promise: nil, requestStreamInitializer)
     }
     return try responseReceivedPromise.futureResult.wait()
 }
@@ -115,21 +114,25 @@ final class HTTP1TestServer: ChannelInboundHandler {
             return
         }
 
+        let loopBoundSelf = NIOLoopBound(self, eventLoop: context.eventLoop)
+        let loopBoundContext = NIOLoopBound(context, eventLoop: context.eventLoop)
+
         // Insert an event loop tick here. This more accurately represents real workloads in SwiftNIO, which will not
         // re-entrantly write their response frames.
+        let channel = context.channel
         context.eventLoop.execute {
-            context.channel.getOption(HTTP2StreamChannelOptions.streamID).flatMap { (streamID) -> EventLoopFuture<Void> in
+            channel.getOption(HTTP2StreamChannelOptions.streamID).flatMap { (streamID) -> EventLoopFuture<Void> in
                 var headers = HTTPHeaders()
                 headers.add(name: "content-length", value: "5")
                 headers.add(name: "x-stream-id", value: String(Int(streamID)))
-                context.channel.write(self.wrapOutboundOut(HTTPServerResponsePart.head(HTTPResponseHead(version: .init(major: 2, minor: 0), status: .ok, headers: headers))), promise: nil)
+                channel.write(loopBoundSelf.value.wrapOutboundOut(HTTPServerResponsePart.head(HTTPResponseHead(version: .init(major: 2, minor: 0), status: .ok, headers: headers))), promise: nil)
 
-                var buffer = context.channel.allocator.buffer(capacity: 12)
+                var buffer = channel.allocator.buffer(capacity: 12)
                 buffer.writeStaticString("hello")
-                context.channel.write(self.wrapOutboundOut(HTTPServerResponsePart.body(.byteBuffer(buffer))), promise: nil)
-                return context.channel.writeAndFlush(self.wrapOutboundOut(HTTPServerResponsePart.end(nil)))
+                channel.write(loopBoundSelf.value.wrapOutboundOut(HTTPServerResponsePart.body(.byteBuffer(buffer))), promise: nil)
+                return channel.writeAndFlush(loopBoundSelf.value.wrapOutboundOut(HTTPServerResponsePart.end(nil)))
             }.whenComplete { _ in
-                context.close(promise: nil)
+                loopBoundContext.value.close(promise: nil)
             }
         }
     }
